Module Tjr_impfs.Blk_v2_UNUSED

Another type for blks; discrimnate more at the type level between differently-sized blks.

For a given type t1, we construct another type t2 which is intended to be the subset of the original; we also construct a singleton type with a value that gives information about the restricted type t2; further, we tie these together by linking the singleton type with t2

type 't2 blk_info = private {

blk_sz : int;

}

module type S = sig ... end

module Make (S : S) : sig ... end

Now we can make a blk type with a given size, freely use it as the underlying type (via coercion) and also use blk_info to get information about the restricted type. However, we can't create new values of this type