Module Tjr_impfs.Dv3

module Flag = Fv3.Flag

module List : sig ... end

val dont_log : bool

type times = Tjr_minifs.Minifs_intf.times

val bin_shape_times : Bin_prot.Shape.t

val bin_size_times : times Bin_prot.Size.sizer

val bin_write_times : times Bin_prot.Write.writer

val bin_writer_times : times Bin_prot.Type_class.writer

val __bin_read_times__ : (int -> times) Bin_prot.Read.reader

val bin_read_times : times Bin_prot.Read.reader

val bin_reader_times : times Bin_prot.Type_class.reader

val bin_times : times Bin_prot.Type_class.t

type ('k, 'v, 't, 'did) dir_ops = {

find : 'k -> ('v option, 't) Tjr_monad.m;

insert : 'k -> 'v -> (unit, 't) Tjr_monad.m;

delete : 'k -> (unit, 't) Tjr_monad.m;

ls_create : unit -> (('k, 'v, 't) Tjr_btree.Btree_intf.ls, 't) Tjr_monad.m;

set_parent : 'did -> (unit, 't) Tjr_monad.m;

get_parent : unit -> ('did, 't) Tjr_monad.m;

set_times : times -> (unit, 't) Tjr_monad.m;

get_times : unit -> (times, 't) Tjr_monad.m;

flush : unit -> (unit, 't) Tjr_monad.m;

sync : unit -> (unit, 't) Tjr_monad.m;

}

type ('did, 'cache) d_cache = {

times : times Fv3.Maybe_dirty.maybe_dirty;

parent : 'did Fv3.Maybe_dirty.maybe_dirty;

cache : 'cache;

}

type 'a cache_entry =

| Some_ of 'a

| Deleted

| Not_present

type ('k, 'v, 'cache) cache_ops = ('k, 'v, 'cache) Tjr_fs_shared.wbc_ops

Make_v2 - assumes lower layer implements batch operations

module Op : sig ... end

module Ls : sig ... end

A potentially nicer interface to list operations?

module Dir : sig ... end

type ('k, 'v, 't, 'did) lower_ops = {

get_meta : unit -> ('did * times, 't) Tjr_monad.m;
returns parent and times

find : 'k -> ('v option, 't) Tjr_monad.m;

exec : ('k, 'v, 't, 'did) Op.exec_ops -> (unit, 't) Tjr_monad.m;

opendir : unit -> (('k, 'v, 't) Ls.ops, 't) Tjr_monad.m;

}

module Make_v2 : functor (S : sig ... end) -> sig ... end

Like Make_v1, but assume that the lower layer supports an exec(ops) operation, rather than individual operations, since we typically flush batches of operations