Module Dv3.Make_v2

Like Make_v1, but assume that the lower layer supports an exec(ops) operation, rather than individual operations, since we typically flush batches of operations

Parameters

S : sig ... end

Signature

val (>>=) : ('a, S.t) Tjr_monad.m -> ('a -> ('b, S.t) Tjr_monad.m) -> ('b, S.t) Tjr_monad.m

val return : 'a -> ('a, S.t) Tjr_monad.m

val get_times : unit -> (times, S.t) Tjr_monad.m

val set_times : times -> (unit, S.t) Tjr_monad.m

val get_parent : unit -> (S.did, S.t) Tjr_monad.m

val set_parent : S.did -> (unit, S.t) Tjr_monad.m

val find : S.k -> (S.v option, S.t) Tjr_monad.m

val lower_insert_sync : (S.k * S.v cache_entry) list -> (unit, S.t) Tjr_monad.m

val maybe_trim : state:('a, 'b) d_cache -> set_state:(('a, S.cache) d_cache -> (unit, S.t) Tjr_monad.m) -> cache:S.cache -> (unit, S.t) Tjr_monad.m

val insert : S.k -> S.v -> (unit, S.t) Tjr_monad.m

val delete : S.k -> (unit, S.t) Tjr_monad.m

val flush : unit -> (unit, S.t) Tjr_monad.m

val sync : unit -> (unit, S.t) Tjr_monad.m

val opendir : unit -> ((S.k, S.v, S.t) Ls.ops, S.t) Tjr_monad.m

val dir_ops : (S.k, S.v, S.t, S.did) Dir.dir_ops