Module Tjr_impfs.Dv3_UNUSED

This is a cache-aware implementation of directories, for v3.

Don't open - too many types with names duplicated elsewhere.

We model a dir as a map from name to dir entry, together with the parent and times.

The cache maps a name to `New _ | `Deleted _ | `Not_present _

Find, insert, delete are straightforward (as are parent and times).

For listing a directory, we take the contents of the cache (held in memory) and combine it with the underlying list operation. The operations for listing are:

type ('k,'v,'r,'ls_state,'t) leaf_stream_ops = {
    make_leaf_stream : 'r -> ('ls_state,'t) m;
    ls_step          : 'ls_state -> ('ls_state option,'t) m;
    ls_kvs           : 'ls_state -> ('k*'v)list;
  }

If these operations are supported by the lower layer, then we just make sure ls_kvs returns the appropriate set of kvs, taking into account the contents of the cache.

module Flag = Fv3_UNUSED.Flag

module List : sig ... end

val dont_log : bool

type times = Tjr_minifs.Minifs_intf.times

val bin_shape_times : Bin_prot.Shape.t

val bin_size_times : times Bin_prot.Size.sizer

val bin_write_times : times Bin_prot.Write.writer

val bin_writer_times : times Bin_prot.Type_class.writer

val __bin_read_times__ : (int -> times) Bin_prot.Read.reader

val bin_read_times : times Bin_prot.Read.reader

val bin_reader_times : times Bin_prot.Type_class.reader

val bin_times : times Bin_prot.Type_class.t

type ('k, 'v, 't, 'did) dir_ops = {

find : 'k -> ('v option, 't) Tjr_monad.m;

insert : 'k -> 'v -> (unit, 't) Tjr_monad.m;

delete : 'k -> (unit, 't) Tjr_monad.m;

ls_create : unit -> (('k, 'v, 't) Tjr_btree.Btree_intf.ls, 't) Tjr_monad.m;

set_parent : 'did -> (unit, 't) Tjr_monad.m;

get_parent : unit -> ('did, 't) Tjr_monad.m;

set_times : times -> (unit, 't) Tjr_monad.m;

get_times : unit -> (times, 't) Tjr_monad.m;

flush : unit -> (unit, 't) Tjr_monad.m;

sync : unit -> (unit, 't) Tjr_monad.m;

}

type ('did, 'cache) d_cache = {

times : times Fv3_UNUSED.Maybe_dirty.maybe_dirty;

parent : 'did Fv3_UNUSED.Maybe_dirty.maybe_dirty;

cache : 'cache;

}

type 'a cache_entry =

| Some_ of 'a

| Deleted

| Not_present

type ('k, 'v, 't) cache_ops = {

find : 'k -> 't -> ('v * bool) option * 't;

insert : 'k -> ('v * bool) -> 't -> 't;

delete : 'k -> 't -> 't;

needs_trim : 't -> bool;

trim : 't -> ('k * 'v) list * 't;

clean : 't -> ('k * 'v) list * 't;

bindings : 't -> ('k * ('v * bool)) list;

}

Make_v2 - assumes lower layer implements batch operations

module Op : sig ... end

module Ls : sig ... end

A potentially nicer interface to list operations?

module Ls_mut : sig ... end

A list interface with mutability

module Dir : sig ... end

type ('k, 'v, 't, 'did) lower_ops = {

get_meta : unit -> ('did * times, 't) Tjr_monad.m;	(*
returns parent and times
*)

find : 'k -> ('v option, 't) Tjr_monad.m;

exec : ('k, 'v, 't, 'did) Op.exec_ops -> (unit, 't) Tjr_monad.m;

opendir : unit -> (('k, 'v, 't) Ls.ops, 't) Tjr_monad.m;

}

module Make_v2 (S : sig ... end) : sig ... end

Like Make_v1, but assume that the lower layer supports an exec(ops) operation, rather than individual operations, since we typically flush batches of operations