Module Dv3_UNUSED.Ls_mut

A list interface with mutability

type ('k, 'v, 'd, 't) ops = {

kvs : 'd -> ('k * 'v) list;	(*
NOTE if we return , then we are not necessarily finished!
*)

step : 'd -> (unit, 't) Tjr_monad.m;

is_finished : 'd -> bool;

close : 'd -> unit;

}

'd is a descriptor; we avoid the monad where possible: the intended implementation is that the "current" kvs is kept in a reference and updated on step. In this case, opendir should return not an ops but a 'd.