Module Dv3_UNUSED.Make_v2

Like Make_v1, but assume that the lower layer supports an exec(ops) operation, rather than individual operations, since we typically flush batches of operations

Parameters

module S : sig ... end

Signature

val (>>=) : ('a, S.t) Tjr_monad.m -> ('a -> ('b, S.t) Tjr_monad.m) -> ('b, S.t) Tjr_monad.m

val return : 'a -> ('a, S.t) Tjr_monad.m

val get_times : unit -> (times, S.t) Tjr_monad.m

val set_times : times -> (unit, S.t) Tjr_monad.m

val get_parent : unit -> (S.did, S.t) Tjr_monad.m

val set_parent : S.did -> (unit, S.t) Tjr_monad.m

val find : S.k -> (S.v option, S.t) Tjr_monad.m

val lower_insert_sync : (S.k * S.v cache_entry) list -> (unit, S.t) Tjr_monad.m

val maybe_trim : state:('a, 'b) d_cache -> set_state:(('a, S.cache) d_cache -> (unit, S.t) Tjr_monad.m) -> cache:S.cache -> (unit, S.t) Tjr_monad.m

val insert : S.k -> S.v -> (unit, S.t) Tjr_monad.m

val delete : S.k -> (unit, S.t) Tjr_monad.m

val flush : unit -> (unit, S.t) Tjr_monad.m

val sync : unit -> (unit, S.t) Tjr_monad.m

The ls operation ("leaf stream") is used to implement readdir. We have a lower uncached implementation, the cache contents, and we need to implement the cached version.

Cache entries are: Some _ | Deleted | Not_present

These entries can be dirty or clean (Not_present is always clean).

A stupid implementation would just flush to lower, and then use the lower operations (since the cache is now clean).

let ls_create () = flush () >>= fun () -> lower_ops.ls_create ()

This is clearly suboptimal, since we shouldn't HAVE to flush to lower, and this operation is potentially costly.

How to implement the cached version of ls? We can ignore Not_present. For deleted cache entries, we must make sure to filter the relevant (k,v) entries from the lower ls.

What to do with new Some entries? There are two obvious implementations:

1) Return all new entries in the first call to ls_kvs(); ls_step() then reverts to the (filtered) entries from lower 2) Attempt to return new entries as we iterate over lower entries via lower.ls_step()

Option 1 is ugly because it reveals the cache contents, and potentially destroys whatever order the lower layer may provide (eg lexicographic on names... not that POSIX requires such an order of course).

Option 2 is somewhat complicated. The issue is where to put new entries that might appear in one of two possible chunks. For example, suppose the first chunk contains keys 1,2,3 and the next (after ls_step()) contains 5,6,7. Where should we put an entry with key 4? To include in the first chunk we need to know that the next chunk does not contain a key 4, which requires reading the next chunk in advance, which is a bit ugly. To include in the second chunk is possible, and avoids the reading ahead. Potentially this requires an extra chunk at the end.

An alternative, if we knew the bounds for a particular chunk (k1 <= ... < k2), would be straightforward - we just include the new entry in the correct chunk.

Ideally we would alter the ls operations to return the bound, and implement the straightforward approach. Rather than go back and alter the ls interface (again!) we hack something together here based on option 1.

val opendir : unit -> ((S.k, S.v, S.t) Ls.ops, S.t) Tjr_monad.m

val dir_ops : (S.k, S.v, S.t, S.did) Dir.dir_ops