Module Dv3_UNUSED.Op

type ('k, 'v, 'did) op' =

| Insert of 'k * 'v

| Delete of 'k

| Set_parent of 'did

| Set_times of times

type exec_type =

| Spill

| With_sync

With the current impl, the spill will also cause the database to update, but perhaps we could consider having another layer of caching here FIXME perhaps we want this just with_sync

type ('k, 'v, 't, 'did) exec_ops = exec_type * ('k, 'v, 'did) op' list