Module Tjr_impfs.File_impl_v1

A simple file implementation as a map from blk_index to blk, typically using a B-tree; NOTE deprecated, do not use.

Terminology: (file) block-index map: the map from blk index to blk_id

We implement a file as a map from blk_index to blk. The map is typically implemented by a B-tree or similar.

The inode contains the file size and the root of the blk_index map.

New blocks are allocated as needed (by default the map is empty, and a block is only allocated when that block is written to; so files are "sparse").

The "iterated block read/write" component has been abstracted out because it appears in a lot of other places.

The components we require are as follows (see "make" function below; NOTE missing free, since this is not currently implemented for file.truncate).

 let make (type fid blk blk_id t) 
      ~monad_ops
      ~(blk_ops       : blk blk_ops)
      ~(blk_dev_ops   : (blk_id,blk,t) blk_dev_ops)
      ~(blk_index_map : (int,blk_id,blk_id,t)pre_map_ops)
      ~(with_inode    : ((fid,blk_id)inode,t)with_state)
      ~(alloc         : unit -> (blk_id,t)m)
  =

Concurrency: The backing B-tree is itself a persistent object.

TODO:

b-tree to implement delete_after (or some alternative approach)

validate against "normal" file semantics

ensure the tests are run by some top-level test executable

implement freeing of blocks after truncate

maybe worth distinguishing the blk_id used by the map from the blk_id used to store data

type 'fid file_id = {

fid : 'fid;

}

module Inode : sig ... end

module Pre_map_ops : sig ... end

Pre_map_ops: because the file root is stored in the inode, we need to access the B-tree using explicit root passing (and then separately update the inode); this is to ensure that the inode update is atomic (via with_inode). We could store the root in a separate block but this is a bit too inefficient.

type pread_error =

| Pread_error of string

type pwrite_error =

| Pwrite_error of string

type ('buf, 't) file_ops = {

size : unit -> (Tjr_fs_shared.Int_like.size, 't) Tjr_monad.m;

pwrite : src:'buf -> src_off:Tjr_fs_shared.Int_like.offset -> src_len:Tjr_fs_shared.Int_like.len -> dst_off:Tjr_fs_shared.Int_like.offset -> ((Tjr_fs_shared.Int_like.size, pwrite_error) Stdlib.result, 't) Tjr_monad.m;

pread : off:Tjr_fs_shared.Int_like.offset -> len:Tjr_fs_shared.Int_like.len -> (('buf, pread_error) Stdlib.result, 't) Tjr_monad.m;

truncate : size:Tjr_fs_shared.Int_like.size -> (unit, 't) Tjr_monad.m;

}

Standard file operations, pwrite, pread, size and truncate.

NOTE we expect buf to be string for the functional version; for mutable buffers we may want to pass the buffer in as a parameter?

NOTE for pwrite, we always return src_len since all bytes are written (unless there is an error of course).

For pread, we always return a buffer of length len.

type ('buf, 'blk, 'blk_id, 'fid, 't) file_factory = < with_ : blk_dev_ops:('blk_id, 'blk, 't) Tjr_fs_shared.blk_dev_ops -> with_inode:(('fid, 'blk_id) Inode.inode, 't) Tjr_monad.with_state -> alloc:(unit -> ('blk_id, 't) Tjr_monad.m) -> < with_blk_index_map : (int, 'blk_id, 'blk_id, 't) Pre_map_ops.pre_map_ops -> ('buf, 't) file_ops; with_btree : blk_id_mshlr:'blk_id Tjr_fs_shared.bp_mshlr -> ('buf, 't) file_ops; >; >

type ('a, 'b) iso = {

a_to_b : 'a -> 'b;

b_to_a : 'b -> 'a;

}

val pread_check : size:int Tjr_fs_shared.Int_like.size_ -> off:int Tjr_fs_shared.Int_like.offset_ -> len:'a Tjr_fs_shared.Int_like.len_ -> (unit, string) Stdlib.result

Check the arguments to pread

val pwrite_check : buf_ops:'buf Tjr_fs_shared.buf_ops -> src:'buf -> src_off:Tjr_fs_shared.Int_like.offset -> src_len:Tjr_fs_shared.Int_like.len -> dst_off:Tjr_fs_shared.Int_like.offset -> (unit, string) Stdlib.result

Check arguments to pwrite