Module File_impl_v1.Pre_map_ops

Pre_map_ops: because the file root is stored in the inode, we need to access the B-tree using explicit root passing (and then separately update the inode); this is to ensure that the inode update is atomic (via with_inode). We could store the root in a separate block but this is a bit too inefficient.

type ('k, 'v, 'r, 't) pre_map_ops = {

find : r:'r -> k:'k -> ('v option, 't) Tjr_monad.m;

insert : r:'r -> k:'k -> v:'v -> ('r option, 't) Tjr_monad.m;

delete : r:'r -> k:'k -> ('r, 't) Tjr_monad.m;

delete_after : r:'r -> k:'k -> ('r, 't) Tjr_monad.m;	(*
delete all entries for keys > r; used for truncate
*)

}