Module Tjr_impfs.File_impl_v2

type 'fid file_id = {

fid : 'fid;

}

module Usedlist : sig ... end

NOTE the in-memory state of the usedlist is opaque to us; after operations, we check the origin to see if it has changed and then possibly flush/barrier/sync

module Freelist : sig ... end

module Pre_map_ops : sig ... end

Pre_map_ops: because the file root is stored in the inode, we need to access the B-tree using explicit root passing (and then separately update the inode); this is to ensure that the inode update is atomic (via with_inode). We could store the root in a separate block but this is a bit too inefficient.

module File_origin_block : sig ... end

We reconstruct the file from the contents of the file's origin blk; this includes the size, the blk->index map root, and the pointers for the used list (which is used to implement alloc in conjunction with the freelist)

module Fo = File_origin_block

module File_im : sig ... end

State we hold in memory for a particular file

type pread_error =

| Pread_error of string

type pwrite_error =

| Pwrite_error of string

type ('buf, 't) file_ops = {

size : unit -> (Tjr_fs_shared.Int_like.size, 't) Tjr_monad.m;

pwrite : src:'buf -> src_off:Tjr_fs_shared.Int_like.offset -> src_len:Tjr_fs_shared.Int_like.len -> dst_off:Tjr_fs_shared.Int_like.offset -> ((Tjr_fs_shared.Int_like.size, pwrite_error) Stdlib.result, 't) Tjr_monad.m;

pread : off:Tjr_fs_shared.Int_like.offset -> len:Tjr_fs_shared.Int_like.len -> (('buf, pread_error) Stdlib.result, 't) Tjr_monad.m;

truncate : size:Tjr_fs_shared.Int_like.size -> (unit, 't) Tjr_monad.m;

flush : unit -> (unit, 't) Tjr_monad.m;

sync : unit -> (unit, 't) Tjr_monad.m;

}

Standard file operations, pwrite, pread, size and truncate.

NOTE we expect buf to be string for the functional version; for mutable buffers we may want to pass the buffer in as a parameter?

NOTE for pwrite, we always return src_len since all bytes are written (unless there is an error of course).

For pread, we always return a buffer of length len.

type ('buf, 'blk, 'blk_id, 't) file_factory = < read_origin : blk_dev_ops:('blk_id, 'blk, 't) Tjr_fs_shared.blk_dev_ops -> blk_id:'blk_id -> ('blk_id File_origin_block.t, 't) Tjr_monad.m; write_origin : blk_dev_ops:('blk_id, 'blk, 't) Tjr_fs_shared.blk_dev_ops -> blk_id:'blk_id -> origin:'blk_id File_origin_block.t -> (unit, 't) Tjr_monad.m; origin_to_fim : 'blk_id File_origin_block.t -> 'blk_id File_im.t; usedlist_origin : 'blk_id File_origin_block.t -> 'blk_id Usedlist.origin; with_ : blk_dev_ops:('blk_id, 'blk, 't) Tjr_fs_shared.blk_dev_ops -> barrier:(unit -> (unit, 't) Tjr_monad.m) -> sync:(unit -> (unit, 't) Tjr_monad.m) -> freelist_ops:('blk_id, 't) Freelist.ops -> < usedlist_ops : 'blk_id Usedlist.origin -> (('blk_id, 't) Usedlist.ops, 't) Tjr_monad.m; alloc_via_usedlist : ('blk_id, 't) Usedlist.ops -> < alloc_via_usedlist : unit -> ('blk_id, 't) Tjr_monad.m; >; blk_index_map_ops : 'blk_id -> (int, 'blk_id, 'blk_id, 't) Pre_map_ops.pre_map_ops; file_ops : usedlist_ops:('blk_id, 't) Usedlist.ops -> alloc_via_usedlist:(unit -> ('blk_id, 't) Tjr_monad.m) -> blk_index_map_ops:(int, 'blk_id, 'blk_id, 't) Pre_map_ops.pre_map_ops -> with_fim:('blk_id File_im.t, 't) Tjr_monad.with_state -> ('buf, 't) file_ops; >; >

type ('a, 'b) iso = {

a_to_b : 'a -> 'b;

b_to_a : 'b -> 'a;

}

val pread_check : size:int Tjr_fs_shared.Int_like.size_ -> off:int Tjr_fs_shared.Int_like.offset_ -> len:'a Tjr_fs_shared.Int_like.len_ -> (unit, string) Stdlib.result

Check the arguments to pread

val pwrite_check : buf_ops:'a Tjr_fs_shared.Buffers_from_btree.buf_ops -> src:'a -> src_off:Tjr_fs_shared.Int_like.offset -> src_len:Tjr_fs_shared.Int_like.len -> dst_off:Tjr_fs_shared.Int_like.offset -> (unit, string) Stdlib.result

Check arguments to pwrite

module Iter_block_blit = File_impl_v1.Iter_block_blit

module type S = sig ... end

module type T = sig ... end

module Make : functor (S : S) -> T with module S = S