Module File_impl_v2.File_origin_block

We reconstruct the file from the contents of the file's origin blk; this includes the size, the blk->index map root, and the pointers for the used list (which is used to implement alloc in conjunction with the freelist)

type 'blk_id file_origin_block = {

file_size : int;

times : stat_times;

blk_idx_map_root : 'blk_id;

usedlist_origin : 'blk_id Usedlist.origin;

}

val bin_shape_file_origin_block : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_file_origin_block : blk_id. 'blk_id Bin_prot.Size.sizer -> 'blk_id file_origin_block Bin_prot.Size.sizer

val bin_write_file_origin_block : blk_id. 'blk_id Bin_prot.Write.writer -> 'blk_id file_origin_block Bin_prot.Write.writer

val bin_writer_file_origin_block : 'a Bin_prot.Type_class.writer -> 'a file_origin_block Bin_prot.Type_class.writer

val __bin_read_file_origin_block__ : blk_id. 'blk_id Bin_prot.Read.reader -> (int -> 'blk_id file_origin_block) Bin_prot.Read.reader

val bin_read_file_origin_block : blk_id. 'blk_id Bin_prot.Read.reader -> 'blk_id file_origin_block Bin_prot.Read.reader

val bin_reader_file_origin_block : 'a Bin_prot.Type_class.reader -> 'a file_origin_block Bin_prot.Type_class.reader

val bin_file_origin_block : 'a Bin_prot.Type_class.t -> 'a file_origin_block Bin_prot.Type_class.t

type 'blk_id t = 'blk_id file_origin_block

val bin_shape_t : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_t : blk_id. 'blk_id Bin_prot.Size.sizer -> 'blk_id t Bin_prot.Size.sizer

val bin_write_t : blk_id. 'blk_id Bin_prot.Write.writer -> 'blk_id t Bin_prot.Write.writer

val bin_writer_t : 'a Bin_prot.Type_class.writer -> 'a t Bin_prot.Type_class.writer

val __bin_read_t__ : blk_id. 'blk_id Bin_prot.Read.reader -> (int -> 'blk_id t) Bin_prot.Read.reader

val bin_read_t : blk_id. 'blk_id Bin_prot.Read.reader -> 'blk_id t Bin_prot.Read.reader

val bin_reader_t : 'a Bin_prot.Type_class.reader -> 'a t Bin_prot.Type_class.reader

val bin_t : 'a Bin_prot.Type_class.t -> 'a t Bin_prot.Type_class.t