Module Make_v1.With_

Parameters

module S2 : sig ... end

Signature

val freelist_ops : (S.blk_id, S.blk_id, S.t) Freelist_intf.freelist_ops

val blk_sz : int

val usedlist_factory' : < alloc_via_usedlist : (S.blk_id, S.t) V2_usedlist_impl.Usedlist.ops -> (S.blk_id, S.t) Tjr_fs_shared.blk_allocator_ops; create : unit -> ((S.blk_id, S.t) V2_usedlist_impl.Usedlist.ops, S.t) Tjr_monad.m; usedlist_ops : S.blk_id V2_usedlist_impl.Usedlist.origin -> ((S.blk_id, S.t) V2_usedlist_impl.Usedlist.ops, S.t) Tjr_monad.m; >

val usedlist_ops : S.blk_id V2_usedlist_impl.Usedlist.origin -> ((S.r, S.t) V2_usedlist_impl.Usedlist.ops, S.t) Tjr_monad.m

Implement the usedlist, using the plist and the global freelist. NOTE For the usedlist, we may hold a free block temporarily; if it isn't used we can return it immediately to the global freelist.

val alloc_via_usedlist : (S.blk_id, S.t) V2_usedlist_impl.Usedlist.ops -> (S.blk_id, S.t) Tjr_fs_shared.blk_allocator_ops

val mk_blk_idx_map : usedlist:(S.blk_id, S.t) V2_usedlist_impl.Usedlist.ops -> btree_root:S.r -> (int, S.r, S.r, S.t) Fv2_types.Btree_ops.t

module File_ops (S3 : sig ... end) : sig ... end

val file_ops : usedlist:(S.blk_id, S.t) V2_usedlist_impl.Usedlist.ops -> alloc_via_usedlist:(unit -> (S.blk_id, S.t) Tjr_monad.m) -> blk_idx_map:(int, S.blk_id, S.blk_id, S.t) Fv2_types.Btree_ops.t -> file_origin:S.blk_id -> init_file_size:int -> init_times:Times.times -> (S.blk_id, S.buf, S.t) file_ops

val create_file : stat_times -> (S.blk_id, S.t) Tjr_monad.m

val file_from_origin_blk : (S.blk_id * S.blk_id File_origin_block.file_origin_block) -> ((S.blk_id, S.buf, S.t) file_ops, S.t) Tjr_monad.m

val file_from_origin : S.blk_id -> ((S.blk_id, S.buf, S.t) file_ops, S.t) Tjr_monad.m

val export : < alloc_via_usedlist : (S.blk_id, S.t) V2_usedlist_impl.Usedlist.ops -> (S.blk_id, S.t) Tjr_fs_shared.blk_allocator_ops; create_file : stat_times -> (S.blk_id, S.t) Tjr_monad.m; file_from_origin : S.blk_id -> ((S.blk_id, S.buf, S.t) file_ops, S.t) Tjr_monad.m; file_from_origin_blk : (S.blk_id * S.blk_id File_origin_block.file_origin_block) -> ((S.blk_id, S.buf, S.t) file_ops, S.t) Tjr_monad.m; file_ops : usedlist:(S.blk_id, S.t) V2_usedlist_impl.Usedlist.ops -> alloc_via_usedlist:(unit -> (S.blk_id, S.t) Tjr_monad.m) -> blk_idx_map:(int, S.blk_id, S.blk_id, S.t) Fv2_types.Btree_ops.t -> file_origin:S.blk_id -> init_file_size:int -> init_times:Times.times -> (S.blk_id, S.buf, S.t) file_ops; mk_blk_idx_map : usedlist:(S.blk_id, S.t) V2_usedlist_impl.Usedlist.ops -> btree_root:S.r -> (int, S.r, S.r, S.t) Fv2_types.Btree_ops.t; usedlist_ops : S.blk_id V2_usedlist_impl.Usedlist.origin -> ((S.r, S.t) V2_usedlist_impl.Usedlist.ops, S.t) Tjr_monad.m; >