Module Tjr_impfs.Fl_make_1

Implement a concurrent-safe freelist using the persistent list.

There are two versions (controlled by a "version" value in the input sig). The first is for arbitrary elements. The second is for blk_ids (ie, a standard freelist).

This module implements the operations without consideration for the origin. At the end of this file, we wrap the operations with a function that updates the origin

module type S = sig ... end

Input sig

module type T = sig ... end

Result sig for Make_v2

module Make_v1 (S : S) : sig ... end

Make with full sig

module Make_v2 (S : S) : T with module S = S

Make with restricted sig

val make : < async : (unit -> (unit, 't) Tjr_monad.m) -> (unit, 't) Tjr_monad.m; event_ops : 't Tjr_monad.event_ops; min_free_alloc : 'elt -> int -> 'elt list * 'elt; monad_ops : 't Tjr_monad.monad_ops; params : Freelist_intf.params; plist_ops : ('elt, 'buf, 'blk_id, 't) Tjr_plist.Plist_intf.plist_ops; sync : unit -> (unit, 't) Tjr_monad.m; version : ('elt, 'blk_id, 't) Freelist_intf.version; with_freelist : ('elt Freelist_intf.freelist_im, 't) Tjr_monad.with_state; .. > -> ('elt, 'blk_id, 't) Freelist_intf.freelist_ops

Make without functor

val wrap_with_origin_updates : monad_ops:'a Tjr_monad.monad_ops -> freelist_ops:('b, 'c, 'a) Freelist_intf.freelist_ops -> write_origin:(fl_origin:('b, 'c) Freelist_intf.Fl_origin.t -> (unit, 'a) Tjr_monad.m) -> ('b, 'c, 'a) Freelist_intf.freelist_ops