Module Fl_make_1.Make_v1

Make with full sig

Parameters

module S : S

Signature

module S = S

type nonrec version = (S.elt, S.blk_id, S.t) Freelist_intf.version

val tr_upper : int

val tr_lower : int

val min_free_alloc_size : int

val (>>=) : ('a, S.t) Tjr_monad.m -> ('a -> ('b, S.t) Tjr_monad.m) -> ('b, S.t) Tjr_monad.m

val return : 'a -> ('a, S.t) Tjr_monad.m

val ev_create : unit -> ('a Tjr_monad.event, S.t) Tjr_monad.m

val ev_wait : 'a Tjr_monad.event -> ('a, S.t) Tjr_monad.m

val ev_signal : 'a Tjr_monad.event -> 'a -> (unit, S.t) Tjr_monad.m

val add : nxt:S.blk_id -> elt:S.elt -> (S.blk_id option, S.t) Tjr_monad.m

val version : [> `For_arbitrary_elts of (unit -> (S.blk_id, S.t) Tjr_monad.m) * (S.blk_id -> (unit, S.t) Tjr_monad.m) * (S.elt -> (unit, S.t) Tjr_monad.m) | `For_blkids of (S.elt -> S.blk_id) * (S.blk_id -> S.elt) ]

The disk_thread asynchronously attempts to add elements to the transient list by reading from the head of the on-disk list and returning new elts.

val disk_thread : unit -> ([> `Finished | `Unfinished ], S.t) Tjr_monad.m

The disk_thread gets triggered when we fall below tr_lower bound. The disk_thread then repeatedly goes to disk and uses the resulting frees to satisfy the waiting threads; the disk_thread only finishes doing this if the number of transients AFTER satisfying the waiting threads is at least tr_lower.

$(CONSIDER()) we may need to throttle if the number of waiters keeps growing, or at least never shrinks to 0; perhaps the disk thread exponentially increases the number of blocks it scans each time

NOTE that when we advance the hd, the old head becomes free

We need to:

(1) read elts from hd, and advance hd

(2) sync the change to hd to the origin block

(3) (arbitrary elts) free old head or (freelist) add hd to list of free elts

(4) for each waiting thread, satisfy the alloc request with one of the new free elts; if we run out of elts, we need to go back to disk (so we need to loop until all waiting threads have been satisfied)

val alloc : unit -> (S.elt, S.t) Tjr_monad.m

The alloc function. We usually satisfy alloc requests from the in-memory list of transient frees. If this becomes low, we go to disk.

Steps involved:

(1) Check if we need to go to disk

(2) Maybe launch an asynchronous disk thread if we are getting low on transient frees

(2.1) If there are no more frees on disk, we can finally resort to using the "min_free_elt" if there is one.

(2.2) We add "min_free_alloc_size"+|waiting| elts to the transient elts

(2.3) And satisfy any waiting requests

(3) Finally, try to get a free element; if there are no transient frees, we add ourselves to the wait list

$(FIXME("not clear that our request will ever be satisfied?"))

val alloc_many : 'a -> 'b

val free_elts : S.elt list -> (unit, S.t) Tjr_monad.m

Free a list of elements, by appending to the on-disk list.

NOTE the following code is rather tricky; if we have an alloc provided, we use this to extend the list; otherwise we are working with elts as blk_ids, so we can use one of the elts as a backing block for the list itself when we need to extend the list

$(CONSIDER("""This is somewhat inefficient because we invoke plist.add multiple times (rather than appending a whole blk of blk_ids)"""))

val free : S.elt -> (unit, S.t) Tjr_monad.m

Free a single elt. This simply adds to the list of transient elements. If the list is then too big, we split it and append half the elements to the on-disk list.

val free_many : 'a -> 'b

val sync : unit -> (unit, S.t) Tjr_monad.m

val get_origin : unit -> ((S.elt, S.blk_id) Freelist_intf.Fl_origin.fl_origin, S.t) Tjr_monad.m

val freelist_ops : (S.elt, S.blk_id, S.t) Freelist_intf.freelist_ops