Module Tjr_impfs.Freelist_intf

Main interfaces used by the freelist

type ('elt, 'blk_id, 't) version =

| For_blkids of ('elt, 'blk_id) for_blk_ids

| For_arbitrary_elts of {

alloc : unit -> ('blk_id, 't) Tjr_monad.m;

free : 'blk_id -> (unit, 't) Tjr_monad.m;

}	(*
For arbitrary elts, we need a way to allocate and free blocks
*)

There are two versions of the freelist:

For_blkids: to implement the "usual" blk freelist, managing on-disk blocks; the elts are actually blk_ids

For_arbitrary_elts: which allocates and frees arbitrary elements; some other freelist of the first type provides the alloc and free blk functionality

and ('elt, 'blk_id) for_blk_ids = {

e2b : 'elt -> 'blk_id;

b2e : 'blk_id -> 'elt;

}

module Fl_origin : sig ... end

type fIXME

type ('a, 'blk_id, 't) freelist_ops = {

alloc : unit -> ('a, 't) Tjr_monad.m;

alloc_many : int -> (fIXME, 't) Tjr_monad.m;

free : 'a -> (unit, 't) Tjr_monad.m;

free_many : fIXME -> (unit, 't) Tjr_monad.m;

get_origin : unit -> (('a, 'blk_id) Fl_origin.t, 't) Tjr_monad.m;

sync : unit -> (unit, 't) Tjr_monad.m;	(*
NOTE the freelist already ensures it is crash safe; this sync is really for tidy shutdown
*)

}

alloc_many: int is the number of blk_ids although this sort-of forces us to unmarshal the whole block; perhaps prefer also storing the number of elements in the block, precisely for this reason; maybe something to $(CONSIDER())

free_many: free an entire list of blk_ids, by appending to this freelist

val freelist_to_blk_allocator : ('a, 'b, 'c) freelist_ops -> ('a, 'c) Tjr_fs_shared.blk_allocator_ops

type 'a min_free_ops = {

min_free_alloc : 'a -> int -> 'a list * 'a;

}

Working with a min_free element, we provide a method that takes the min_free element, and a count of the number of frees required, and returns a list of free elements (of the same length as requested) and the new min_free

type 'a freelist_im = {

transient : 'a list;

min_free : 'a option;

waiting : 'a Tjr_monad.event list;

disk_thread_active : bool;

}

In-memory state for the freelist

The freelist state, in addition to the plist.

transient: a non-persistent list of free elts that can be allocated

min_free: free elts are stored in transient, or in the on-disk plist; usually we also have an "upper bound", beyond which we can still allocate elts freely (an elt is either in transient, allocated to some object, in the plist, or geq min_free)

waiting: a list of threads waiting for a disk process to return

NOTE: we assume that the state is accessed via with_state (ie with a mutex), so only one thread at a time modifies state

val empty_freelist : min_free:'a option -> 'a freelist_im

type params = < tr_upper : int; tr_lower : int; min_free_alloc_size : int; >

type ('a, 'buf, 'blk_id, 'blk, 't) freelist_factory = < version : ('a, 'blk_id) for_blk_ids; empty_freelist : min_free:'a option -> 'a freelist_im; read_origin : blk_dev_ops:('blk_id, 'blk, 't) Tjr_fs_shared.blk_dev_ops -> blk_id:'blk_id -> (('a, 'blk_id) Fl_origin.t, 't) Tjr_monad.m; write_origin : blk_dev_ops:('blk_id, 'blk, 't) Tjr_fs_shared.blk_dev_ops -> blk_id:'blk_id -> origin:('a, 'blk_id) Fl_origin.t -> (unit, 't) Tjr_monad.m; fl_origin_to_pl : ('a, 'blk_id) Fl_origin.t -> 'blk_id Tjr_plist.Plist_intf.Pl_origin.t; with_ : blk_dev_ops:('blk_id, 'blk, 't) Tjr_fs_shared.blk_dev_ops -> barrier:(unit -> (unit, 't) Tjr_monad.m) -> sync:(unit -> (unit, 't) Tjr_monad.m) -> params:params -> < plist_ops : 'a Tjr_plist.Plist_intf.Pl_origin.t -> (('a, 'buf, 'blk_id, 't) Tjr_plist.Plist_intf.plist_ops, 't) Tjr_monad.m; with_plist_ops : ('a, 'buf, 'blk_id, 't) Tjr_plist.Plist_intf.plist_ops -> < with_state : ('a freelist_im, 't) Tjr_monad.with_state -> ('a, 'blk_id, 't) freelist_ops; with_locked_ref : 'a freelist_im -> < freelist_ops : ('a, 'blk_id, 't) freelist_ops; freelist_ref : 'a freelist_im Stdlib.ref; >; >; add_origin_autosync : origin_blk_id:'blk_id -> ('a, 'blk_id, 't) freelist_ops -> ('a, 'blk_id, 't) freelist_ops; initialize : origin:'blk_id -> free_blk:'blk_id -> min_free:'a option -> (unit, 't) Tjr_monad.m; restore : autosync:bool -> origin:'blk_id -> (< freelist_ops : ('a, 'blk_id, 't) freelist_ops; freelist_ref : 'a freelist_im Stdlib.ref; >, 't) Tjr_monad.m; >; >