Module Tjr_impfs.Fv2_iter_block_blit

Iterated block blit, for file_impl_v2

Abstract model of blitting from blks to buffer.

NOTE The pread,pwrite functions that result will lock the file whilst executing, in order to udpate the B-tree root

Abstract model of blitting from blks to buffer.

Spec: a seq of blits equiv to a single blit, but at most 1 blit is not block aligned, and at most 2 blits are not blk sized; also, minimal number of blits (no overlapping blits).

This version assumes that the file is backed by a map from blk_index -> blk_id (although for the test code, we ignore blk_id and just work directly with bytes.)

read_blk: given a block index, we look up the actual blk_id in the map, and read that block; return empty blk if no blk; ignore size

alloc_and_write_blk: for the situation where we definitely want to allocate a new blk; we have a blk_index, a blk, and we allocate a blk_id, write the blk, insert the (index,blk_id) into the map, then return the blk_id

rewrite_blk_or_alloc: given a blk_index and a blk_id (of the previous version of the blk) and the new blk, we attempt to rewrite that blk directly; if for some reason we cannot mutate the old blk, we allocate a new blk_id, write the new blk, insert the (index,blk_id) into the map

truncate n: drop the blk-index map entries for all blocks after byte index n

For the usage with file_impl, we assume the inode is locked (so we can update the inode using set_state).

type ('buf, 't) ops = {

pwrite : src:'buf -> src_off:Tjr_fs_shared.Int_like.offset -> src_len:Tjr_fs_shared.Int_like.len -> dst_off:Tjr_fs_shared.Int_like.offset -> (Tjr_fs_shared.Int_like.size, 't) Tjr_monad.m;

pread : off:Tjr_fs_shared.Int_like.offset -> len:Tjr_fs_shared.Int_like.len -> ('buf, 't) Tjr_monad.m;

}

module Idx : sig ... end

indexes, unboxed

val idx_mshlr : Idx.t Tjr_fs_shared.bp_mshlr

type idx = Idx.t

module type S = sig ... end

module type T = sig ... end

module Make_v1 (S : S) : sig ... end

Make with full sig

module Make_v2 (S : S) : T with module S = S

Restricted sig

val make : monad_ops:'t Tjr_monad.monad_ops -> buf_ops:'buf Tjr_fs_shared.buf_ops -> blk_ops:('blk, 'buf) Tjr_fs_shared.blk_ops -> read_blk:(idx -> ('blk_id * 'blk, 't) Tjr_monad.m) -> alloc_and_write_blk:(idx -> 'blk -> (unit, 't) Tjr_monad.m) -> rewrite_blk_or_alloc:(idx -> ('blk_id * 'blk) -> (unit, 't) Tjr_monad.m) -> ('buf, 't) ops

As a function, not a functor

val blk_ops_by_by_2 : (bytes, bytes) Tjr_fs_shared.Blk_ops.blk_ops

val test : unit -> unit