Module Fv2_iter_block_blit.Make_v1

Make with full sig

Parameters

module S : S

Signature

module S = S

NOTE we want make to be (as far as possible) independent of the updates to the blk-index map. NOTE that blk_id is essentially completely abstract here, and only used when rewriting.

val (>>=) : ('a, S.t) Tjr_monad.m -> ('a -> ('b, S.t) Tjr_monad.m) -> ('b, S.t) Tjr_monad.m

val return : 'a -> ('a, S.t) Tjr_monad.m

val blk_sz : int

val pread : off:int Tjr_fs_shared.Int_like.offset_ -> len:'a Tjr_fs_shared.Int_like.len_ -> (S.buf, S.t) Tjr_monad.m

This version assumes that the arguments pass the pread_check. We create a buffer of size len.

val pwrite : src:S.buf -> src_off:int Tjr_fs_shared.Int_like.offset_ -> src_len:'a Tjr_fs_shared.Int_like.len_ -> dst_off:int Tjr_fs_shared.Int_like.offset_ -> (int Tjr_fs_shared.Int_like.size_, S.t) Tjr_monad.m

val ops : (S.buf, S.t) ops