Module type Fv2_iter_block_blit.S

type buf

type blk_id

type blk

type t

val monad_ops : t Tjr_monad.monad_ops

val buf_ops : buf Tjr_fs_shared.buf_ops

val blk_ops : (blk, buf) Tjr_fs_shared.blk_ops

val read_blk : idx -> (blk_id * blk, t) Tjr_monad.m

Read a blk, given the index, and some map from idx to blk-id; returns the underlying blk-id, in case we need to rewrite

NOTE blk_id is used only when rewriting

val alloc_and_write_blk : idx -> blk -> (unit, t) Tjr_monad.m

If there is no existing blk for a given index, we need to allocate one and write it (and update the B-tree)

val rewrite_blk_or_alloc : idx -> (blk_id * blk) -> (unit, t) Tjr_monad.m

Try to rewrite a blk; if for some reason (eg snapshot) we can't rewrite, then just allocate another blkid and use that instead (updating the B-tree of course).