Module Fv2_types.File_im

State we hold in memory for a particular file

type 'blk_id t = {

origin_info : 'blk_id File_origin_block.t;

origin_dirty : bool;

}

origin_dirty - whether we need to write out the origin (ie any of the relevant fields have changed)