Module Tjr_impfs.Sqlite_dir

include module type of struct include V3_sqlite.Sqlite_dir end

val assert_ok : Sqlite3.Rc.t -> unit

module Times = Tjr_fs_shared.Times

type times = Times.times

module List : sig ... end

module Ls : sig ... end

module Op : sig ... end

type ('k, 'v, 't, 'did) dir_ops = {

get_meta : did:'did -> ('did * times, 't) Tjr_monad.m;

find : did:'did -> 'k -> ('v option, 't) Tjr_monad.m;

exec : ('k, 'v, 'did) Op.op list -> (unit, 't) Tjr_monad.m;

opendir : did:'did -> (('k, 'v, 't) Ls.ops, 't) Tjr_monad.m;

max_did : unit -> 'did;	(*
NOTE this is blocking, but is intended to be used once at startup, so potentially only results in slightly slower startup than would be possible if this was in the monad
*)

pre_create : note_does_not_touch_parent:unit -> new_did:'did -> parent:'did -> times:times -> (unit, 't) Tjr_monad.m;	(*
pre-create does not touch the parent directory - that needs to be done with a separate insert
*)

}

type db = Sqlite3.db

val create_tables_stmt : string

SQL commands to create a db; dir_entries and dir_meta; NOTE name is 256 length

val create_tables : Sqlite3.db -> unit

val add_root_directory_stmt : string

Add root directory to db

val add_root_directory : Sqlite3.db -> unit

val readdir_limit : int

module Make (S : sig ... end) : sig ... end

module Test () : sig ... end