Module Tjr_impfs.Test_blk_dev

A test blk dev which keeps track of which writes occurred when and can write all intermediate states to a directory of files (numbered 0...)

module M : sig ... end

type map = Bigstring.t M.t

type trace = {

mutable writes : (Tjr_fs_shared.Shared_ctxt.r * string) list;

mutable map : map;

mutable dirty : bool;

}

A list of block writes

val empty_trace : unit -> trace

val blk_sz' : int

val marshal_fn : dirname:string -> string

val max_ : unit -> int

val make : unit -> < blk_dev_ops : (Tjr_fs_shared.Shared_ctxt.r, Bigstring.t, Tjr_monad.With_lwt.lwt) Tjr_fs_shared.blk_dev_ops; close : unit -> (unit, Tjr_monad.With_lwt.lwt) Tjr_monad.m; t0 : trace; write_versions_to_dir : unit -> unit; >

val restore : unit -> < blk_dev_ops : (Tjr_fs_shared.Shared_ctxt.r, Bigstring.t, Tjr_monad.With_lwt.lwt) Tjr_fs_shared.blk_dev_ops; close : unit -> (unit, Tjr_monad.With_lwt.lwt) Tjr_monad.m; t0 : trace; write_versions_to_dir : unit -> unit; >

val print_trace : dirname:string -> unit