Module Make_v1.With_

Parameters

S2 : sig ... end

Signature

val blk_allocator : (S.blk_id, S.t) Tjr_fs_shared.blk_allocator_ops

val blk_sz : int

val plist_factory' : < add_origin : < set_and_sync : S.blk_id Tjr_plist.Plist_intf.Pl_origin.pl_origin -> (unit, S.t) Tjr_monad.m; > -> (S.a, S.buf, S.blk_id, S.t) Tjr_plist.Plist_intf.plist_ops -> (S.a, S.buf, S.blk_id, S.t) Tjr_plist.Plist_intf.plist_ops; init : < create : S.blk_id -> ((S.blk_id, S.buf) Tjr_plist.Plist_intf.plist, S.t) Tjr_monad.m; from_endpts : S.blk_id Tjr_plist.Plist_intf.Pl_origin.pl_origin -> ((S.blk_id, S.buf) Tjr_plist.Plist_intf.plist, S.t) Tjr_monad.m; read_from_hd : S.blk_id -> ((S.a list * S.blk_id option) list, S.t) Tjr_monad.m; >; with_ref : (S.blk_id, S.buf) Tjr_plist.Plist_intf.plist -> < plist_ref : (S.blk_id, S.buf) Tjr_plist.Plist_intf.plist Stdlib.ref; with_plist : ((S.blk_id, S.buf) Tjr_plist.Plist_intf.plist, S.t) Tjr_monad.with_state; >; with_state : ((S.blk_id, S.buf) Tjr_plist.Plist_intf.plist, S.t) Tjr_monad.with_state -> (S.a, S.buf, S.blk_id, S.t) Tjr_plist.Plist_intf.plist_ops; >

val usedlist_ops : S.blk_id Usedlist.origin -> ((S.r, S.t) Usedlist.ops, S.t) Tjr_monad.m

Implement the usedlist, using the plist and the global freelist. NOTE For the usedlist, we may hold a free block temporarily; if it isn't used we can return it immediately to the global freelist.

val alloc_via_usedlist : (S.blk_id, S.t) Usedlist.ops -> (S.blk_id, S.t) Tjr_fs_shared.blk_allocator_ops

val create : unit -> ((S.r, S.t) Usedlist.ops, S.t) Tjr_monad.m

val export : < alloc_via_usedlist : (S.blk_id, S.t) Usedlist.ops -> (S.blk_id, S.t) Tjr_fs_shared.blk_allocator_ops; create : unit -> ((S.r, S.t) Usedlist.ops, S.t) Tjr_monad.m; usedlist_ops : S.blk_id Usedlist.origin -> ((S.r, S.t) Usedlist.ops, S.t) Tjr_monad.m; >