Module Tjr_impfs.V1_generic

An abstract generic development; this layer deals with path resolution, times and a few other things; assumes create_dir and create_file.

type exn_ = Tjr_minifs.Minifs_intf.exn_

type stat_times = Tjr_minifs.Minifs_intf.times

type stat_record = Tjr_minifs.Minifs_intf.stat_record

module File_ops : sig ... end

include module type of struct include File_ops end

type buf = Tjr_fs_shared.Shared_ctxt.buf

type 't file_ops = 't File_ops.file_ops = {

pread : foff:int -> len:int -> buf:buf -> boff:int -> ((int, Tjr_minifs.Minifs_intf.Call_specific_errors.pread_err) Stdlib.result, 't) Tjr_monad.m;

pwrite : foff:int -> len:int -> buf:buf -> boff:int -> ((int, Tjr_minifs.Minifs_intf.Call_specific_errors.pwrite_err) Stdlib.result, 't) Tjr_monad.m;

truncate : int -> (unit, 't) Tjr_monad.m;

get_sz : unit -> (int, 't) Tjr_monad.m;

set_times : stat_times -> (unit, 't) Tjr_monad.m;

get_times : unit -> (stat_times, 't) Tjr_monad.m;

}

module type S0 = sig ... end

Very basic types

module S1 (S0 : S0) : sig ... end

Types created from basic types

module S2 (S0 : S0) : sig ... end

The values we expect to be present

module type T = sig ... end

The resulting filesystem (fd is a file identifier, dh is a directory handle supporting leaf_stream operations)

module Make (S0 : S0) : sig ... end

module X = Make