Module Make.S1

type dir_entry = S1(S0).dir_entry =

| Fid of S0.fid

| Did of S0.did

| Sid of S0.sid

val bin_shape_dir_entry : Bin_prot.Shape.t

val bin_size_dir_entry : dir_entry Bin_prot.Size.sizer

val bin_write_dir_entry : dir_entry Bin_prot.Write.writer

val bin_writer_dir_entry : dir_entry Bin_prot.Type_class.writer

val __bin_read_dir_entry__ : (int -> dir_entry) Bin_prot.Read.reader

val bin_read_dir_entry : dir_entry Bin_prot.Read.reader

val bin_reader_dir_entry : dir_entry Bin_prot.Type_class.reader

val bin_dir_entry : dir_entry Bin_prot.Type_class.t

type dir_k = Tjr_fs_shared.str_256

type dir_v = dir_entry

type dh = (dir_k, dir_v, S0.t) Tjr_btree.Btree_intf.ls

dir handle

type dir_ops = S1(S0).dir_ops = {

find : Tjr_fs_shared.str_256 -> (dir_entry option, S0.t) Tjr_monad.m;

insert : Tjr_fs_shared.str_256 -> dir_entry -> (unit, S0.t) Tjr_monad.m;

delete : Tjr_fs_shared.str_256 -> (unit, S0.t) Tjr_monad.m;

ls_create : unit -> (dh, S0.t) Tjr_monad.m;

set_parent : S0.did -> (unit, S0.t) Tjr_monad.m;

get_parent : unit -> (S0.did, S0.t) Tjr_monad.m;

set_times : stat_times -> (unit, S0.t) Tjr_monad.m;

get_times : unit -> (stat_times, S0.t) Tjr_monad.m;

}

type dirs_ops = S1(S0).dirs_ops = {

find : S0.did -> (dir_ops, S0.t) Tjr_monad.m;

delete : S0.did -> (unit, S0.t) Tjr_monad.m;

}

NOTE for create, use create_dir; this needs to add the did to the gom

type fd = S0.fid

type buf = Tjr_minifs.Minifs_intf.ba_buf

type nonrec file_ops = S0.t file_ops

type files_ops = S1(S0).files_ops = {

find : S0.fid -> (file_ops, S0.t) Tjr_monad.m;

create : S0.fid -> stat_times -> (unit, S0.t) Tjr_monad.m;

}

NOTE create just creates a new file in the gom; it doesn't link it into a parent etc

type resolved_path_or_err = (S0.fid, S0.did, S0.sid) Tjr_path_resolution.resolved_path_or_err

type path = string

type extra_ops = S1(S0).extra_ops = {

internal_err : a. string -> ('a, S0.t) Tjr_monad.m;

is_ancestor : parent:S0.did -> child:S0.did -> (bool, S0.t) Tjr_monad.m;

}