Module V1_resource_manager.Per_object_data

type file_cache

type fid

type did

type sid

type blk_id

type lock = Tjr_fs_shared.Shared_ctxt.mutex

type 'a maybe_dirty = 'a * bool

type unlocked_entry =

| Gom of {

btree_root : blk_id;

cache : unit;

}

| File of {

fid : fid;

origin : blk_id;

times : Tjr_minifs.Minifs_intf.Times.times maybe_dirty;

fd : Unix.file_descr;

}

| Dir of {

did : did;

origin : blk_id;

parent : did maybe_dirty;

btree_root : blk_id maybe_dirty;

times : Tjr_minifs.Minifs_intf.Times.times maybe_dirty;

cache : unit;

}

| Symlink of {

sid : sid;

origin : blk_id;

contents : Tjr_fs_shared.str_256 maybe_dirty;

}

type entry = unlocked_entry * lock

type t = {

entries : entry Map_oid.t;

super_lock : lock;
this has to be held if a thread wants to take a particular object lock; this is so that we can pause all thread lock operations on objects, in order to GC without worrying about other threads trying to take locks

}