Module Tjr_impfs.V1_specific

V1 - implement directories and file metadata, with file data passed through to an underlying filesystem.

We have a single "global object map", which maps ids (ints) to either dir-roots, or file-meta, or symlinks.

NOTE much of this is based on Tjr_minifs.In_mem, but with on-disk maps, rather than in-memory.

module G = V1_generic

Some simple defns we can complete here

val root_did : V1_types.S0.did

val mk_stat_times : unit -> (Tjr_minifs.Minifs_intf.Times.times, Tjr_monad.With_lwt.lwt) Tjr_monad.m

Implement definitions from T2, ie, the core filesystem

type dir_entry = V1_types.S1.dir_entry

Various marshallers

val shared_mshlrs : < ba_mshlr : a. mshlr:'a Tjr_fs_shared.Pvt_bin_prot_marshalling.bp_mshlr -> buf_sz:int -> 'a Tjr_fs_shared.Pvt_bin_prot_marshalling.ba_mshlr; int_mshlr : int Tjr_fs_shared.Pvt_bin_prot_marshalling.bp_mshlr; r_mshlr : Tjr_fs_shared__.Shared_ctxt.r Tjr_fs_shared.Pvt_bin_prot_marshalling.bp_mshlr; s256_mshlr : Tjr_fs_shared__.Str_256.str_256 Tjr_fs_shared.Pvt_bin_prot_marshalling.bp_mshlr; >

val dir_entry_mshlr : dir_entry Tjr_fs_shared.bp_mshlr

The global object map (GOM)

module Gom : sig ... end

Could just use disjoint subsets of int and drop the constructors

module Gom_btree : sig ... end

val gom_factory : (Gom_btree.M1.k, Gom_btree.M1.v, Gom_btree.M1.r, Gom_btree.M1.t, Gom_btree.M1.leaf, Gom_btree.M1.node, (Gom_btree.node, Gom_btree.leaf) Isa_btree.Isa_btree_intf.dnode, Gom_btree.M1.leaf_stream, Tjr_fs_shared.Shared_ctxt.blk, Gom_btree.W_.wbc) Tjr_btree__Make_6.btree_factory

Stage 1 functor

module Stage_1 (R1 : sig ... end) : sig ... end