Module V1_specific.Gom

Could just use disjoint subsets of int and drop the constructors

type id = dir_entry

val bin_shape_id : Bin_prot.Shape.t

val bin_size_id : id Bin_prot.Size.sizer

val bin_write_id : id Bin_prot.Write.writer

val bin_writer_id : id Bin_prot.Type_class.writer

val __bin_read_id__ : (int -> id) Bin_prot.Read.reader

val bin_read_id : id Bin_prot.Read.reader

val bin_reader_id : id Bin_prot.Type_class.reader

val bin_id : id Bin_prot.Type_class.t

val v_size : int

val id_mshlr : id Tjr_fs_shared.bp_mshlr

val consistent_entry : V1_types.S1.dir_entry -> V1_types.S1.dir_entry -> bool

type k = id

val bin_shape_k : Bin_prot.Shape.t

val bin_size_k : k Bin_prot.Size.sizer

val bin_write_k : k Bin_prot.Write.writer

val bin_writer_k : k Bin_prot.Type_class.writer

val __bin_read_k__ : (int -> k) Bin_prot.Read.reader

val bin_read_k : k Bin_prot.Read.reader

val bin_reader_k : k Bin_prot.Type_class.reader

val bin_k : k Bin_prot.Type_class.t

val k_mshlr : k Tjr_fs_shared.bp_mshlr

type v = Tjr_fs_shared.Shared_ctxt.blk_id

The gom maps an id to a root blk

val v_mshlr : v Tjr_fs_shared.bp_mshlr

val k_cmp : id -> id -> int

type r = Tjr_fs_shared.Shared_ctxt.r

val r_cmp : Tjr_fs_shared.Shared_ctxt.r -> Tjr_fs_shared.Shared_ctxt.r -> int

val r_mshlr : Tjr_fs_shared__.Shared_ctxt.r Tjr_fs_shared.Pvt_bin_prot_marshalling.bp_mshlr

val cs : Isa_btree.Constants.constants

type t = Tjr_fs_shared.Shared_ctxt.t

val monad_ops : Tjr_monad.With_lwt.lwt Tjr_monad__.Monad_intf.monad_ops