Module V1_specific.Stage_1

Parameters

module R1 : sig ... end

Signature

val min_free_id_ref : int Stdlib.ref

We use mutually exclusive subsets of int for identifiers; NOTE 0 is reserved for the root directory

val new_id : unit -> (int, Tjr_monad.With_lwt.lwt) Tjr_monad.m

val new_did : unit -> (V1_types.S0.did, Tjr_monad.With_lwt.lwt) Tjr_monad.m

val new_fid : unit -> (V1_types.S0.fid, Tjr_monad.With_lwt.lwt) Tjr_monad.m

val new_sid : unit -> (V1_types.S0.sid, Tjr_monad.With_lwt.lwt) Tjr_monad.m

val id_to_int : V1_types.S1.dir_entry -> int

Definitions after the GOM root is known

val gom_btree : < get_btree_root : unit -> (Gom_btree.M1.r, Gom_btree.M1.t) Tjr_monad.m; map_ops_with_ls : (Gom_btree.M1.k, Gom_btree.M1.v, Gom_btree.M1.r, Gom_btree.M1.leaf_stream, Gom_btree.M1.t) Tjr_btree__.Btree_intf.map_ops_with_ls; with_state : (Gom_btree.M1.r, Gom_btree.M1.t) Tjr_monad.with_state; >

val gom_find_opt : k:Gom_btree.M1.k -> (Gom_btree.M1.v option, Gom_btree.M1.t) Tjr_monad.m

val gom_find : Gom_btree.M1.k -> (Gom_btree.M1.v, Tjr_monad.With_lwt.lwt) Tjr_monad.m

val gom_insert : k:Gom_btree.M1.k -> v:Gom_btree.M1.v -> (unit, Gom_btree.M1.t) Tjr_monad.m

val gom_delete : k:Gom_btree.M1.k -> (unit, Gom_btree.M1.t) Tjr_monad.m

In order to incorporate path resolution, we need to be able to lookup an entry in a directory

module Make_root_blk_ops (S : sig ... end) : sig ... end

module Dir : sig ... end

val resolve_path : follow_last_symlink:Tjr_path_resolution.follow_last_symlink -> Tjr_path_resolution.comp_ -> ((V1_types.S0.fid, V1_types.S0.did, V1_types.S0.sid) Tjr_path_resolution.resolved_path_or_err, Tjr_monad.With_lwt.lwt) Tjr_monad.m

Dirs collection

val dirs : V1_types.S1.dirs_ops

The dirs ops, find and delete FIXME these need to be maintained in a pool, and only one instance per id

File struct

module File : sig ... end

Files collection

module Files : sig ... end

val files : V1_types.S1.files_ops

val extra : V1_types.S1.extra_ops

Extra ops

create_dir, create_file and create_symlink

val create_file : parent:Tjr_impfs__V1_types.S0.did -> name:Tjr_fs_shared.str_256 -> times:Tjr_minifs.Minifs_intf.times -> (unit, Tjr_monad.With_lwt.lwt) Tjr_monad.m

val create_dir_ : is_root:bool -> parent:V1_types.S0.did -> name:Tjr_fs_shared.Str_256.str_256 -> times:Tjr_minifs.Minifs_intf.Times.times -> (unit, Tjr_monad.With_lwt.lwt) Tjr_monad.m

val create_dir : parent:V1_types.S0.did -> name:Tjr_fs_shared.Str_256.str_256 -> times:Tjr_minifs.Minifs_intf.Times.times -> (unit, Tjr_monad.With_lwt.lwt) Tjr_monad.m

val create_root_dir : times:Tjr_minifs.Minifs_intf.Times.times -> (unit, Tjr_monad.With_lwt.lwt) Tjr_monad.m

Symlinks

module Symlink : sig ... end

val create_symlink : parent:Tjr_impfs__V1_types.S0.did -> name:Tjr_fs_shared.str_256 -> times:'a -> contents:Tjr_fs_shared.Str_256.str_256 -> (unit, Tjr_monad.With_lwt.lwt) Tjr_monad.m

val read_symlink : Tjr_impfs__V1_types.S0.sid -> (Tjr_fs_shared.Str_256.str_256, Tjr_monad.With_lwt.lwt) Tjr_monad.m

Instantiate filesystem

module X : sig ... end

module The_filesystem : sig ... end