Module V2.Stage_1

Parameters

module S1 : sig ... end

Signature

val barrier : unit -> (unit, Tjr_monad.With_lwt.lwt) Tjr_monad.m

val sync : unit -> (unit, Tjr_monad.With_lwt.lwt) Tjr_monad.m

New identifiers

We use mutually exclusive subsets of int for identifiers; NOTE 0 is reserved for the root directory

val counter_factory : (Tjr_fs_shared.Shared_ctxt.B.blk_id, Tjr_fs_shared.Shared_ctxt.blk, Tjr_monad.With_lwt.lwt) V2_counter.counter_factory

val counter_ops : (Tjr_monad.With_lwt.lwt V2_counter.counter_ops, Tjr_monad.With_lwt.lwt) Tjr_monad.m

The freelist

We need to read the freelist origin block, and resurrect the freelist

val fl_params : < min_free_alloc_size : int; tr_lower : int; tr_upper : int; >

val freelist : (< freelist_ops : (Tjr_fs_shared.Shared_ctxt.r, Tjr_fs_shared.Shared_ctxt.r, Tjr_monad.With_lwt.lwt) Freelist_intf.freelist_ops; freelist_ref : Tjr_fs_shared.Shared_ctxt.r Freelist_intf.freelist_im Stdlib.ref; >, Tjr_monad.With_lwt.lwt) Tjr_monad.m

val fl_ops : ((Tjr_fs_shared.Shared_ctxt.r, Tjr_fs_shared.Shared_ctxt.r, Tjr_monad.With_lwt.lwt) Freelist_intf.freelist_ops, Tjr_monad.With_lwt.lwt) Tjr_monad.m

The global object map (GOM)

val gom_ops : (V2_gom.Pvt.S.gom_ops, Tjr_monad.With_lwt.lwt) Tjr_monad.m

Stage_2: we assume freelist_ops, gom_ops and counter_ops are available

module Stage_2 (S2 : sig ... end) : sig ... end

val make : unit -> ((S0.fid, (Tjr_fs_shared.str_256, Make.S1.dir_entry, S0.t) Tjr_btree.Btree_intf.ls, S0.t) Tjr_minifs.Minifs_intf.ops, S0.t) Tjr_monad.m