Module Tjr_impfs.V2_generic

V2 generic development based on V1 generic; this layer deals with path resolution, times and a few other things; assumes create_dir and create_file.

$(INDEX("time,times,timestamps"))

Short note on times:

POSIX has an informal spec. SibylFS has a formal spec of timestamps, but here I want a very short description of atime, mtime and ctime, suitable for a rough implementation of times.

Sources:

https://www.howtogeek.com/517098/linux-file-timestamps-explained-atime-mtime-and-ctime/

https://www.unixtutorial.org/atime-ctime-mtime-in-unix-filesystems/

atime: last access time (read? or read/write? what is the meaning of "access"? probably "read of file data") for file or directory; often disabled (since updating causes a disk write)

ctime: last (metadata) change (ownership, access perms, ...); also updated if contents change; probably doesn't affect us since we don't currently have ownership or access perms

mtime: last change to file contents; does not change when eg ownership/perms change

Since "file size" is typically counted as metadata (!), it is clear that parts of metadata are considered differently from others (ctime identifies some subset of file metadata... ctime doesn't change when file size changes I expect).

Personally, I would consider file data (and the related file size) as independent of anything else about the file (ownership, permissions, timestamp). Then it makes sense to talk about when the file data (not the metadata!) was last read, and when it was last written. So, for us, let's define atime as "last read time", ignore ctime (identify with mtime), and mtime as the time the data was last modified (including truncate of course).

Since atime is usually ignored/disabled anyway, we end up with just mtime. For files, this likely changes on pwrite and truncate (set_times confusingly doesn't itself automatically update mtime, since the purpose of set_times is precisely to allow updating the mtime to some specific value).

For directories, we should update mtime on insert and delete. The slight issue is that we may want a cross-directory rename to appear to happen at a single time. For a cross directory rename, we should call "set_times" on both the source and destination directory, with the same timestamp, AFTER the rename has inserted and deleted.

Should files/directories be responsible for updating their own timestamps? As above, this makes sense, except in cases like rename where we may want the src and target dirs to have identical mtimes. (And even in this case, does anyone program applications to work only if this behaviour is implemented by the filesystem?). So we DON'T explicitly update times in this generic code (unless eg for rename case)

type exn_ = Tjr_minifs.Minifs_intf.exn_

type stat_times = Tjr_minifs.Minifs_intf.times

type stat_record = Tjr_minifs.Minifs_intf.stat_record

module type S0 = sig ... end

Very basic types

module S1 (S0 : S0) : sig ... end

Types created from basic types

module S2 (S0 : S0) : sig ... end

The values we expect to be present

module type T = sig ... end

The resulting filesystem (fd is a file identifier, dh is a directory handle supporting leaf_stream operations)

module Make (S0 : S0) : sig ... end

module X = Make