Module Tjr_impfs.V2_live_object_cache

An LRU cache of live objects; used for files and dirs.

This code is a bit naive - it blocks when going to the lower layer, whereas a cleverer version might allow read-only access to in-cache entries whilst going to lower layer. Whether this is actually a bottleneck in practice I don't know.

type ('k, 'v, 'c, 't) factory = < empty : capacity:int -> 'c; with_ : lower_acquire:('k -> ('v option, 't) Tjr_monad.m) -> lower_release:(('k * 'v) -> (unit, 't) Tjr_monad.m) -> with_locked_state:('c, 't) Tjr_monad.with_state -> < find_opt : 'k -> ('v option, 't) Tjr_monad.m; to_list : unit -> (('k * 'v) list, 't) Tjr_monad.m; >; >

factory type; NOTE 'c tyvar is short for 'cache

NOTE lower_acquire is used to look up k in the lower map; lower_release is used when dropping a LRU entry from the cache

module type S = sig ... end

module type T = sig ... end

module Make_v1 (S : S) : sig ... end

With full sig

module Make_v2 (S : S) : T with module S = S

Version with restricted sig

module Make = Make_v2

module Examples : sig ... end

val examples : < for_int_int : (Examples.Int_int.Make'.S.k, Examples.Int_int.Make'.S.v, Examples.Int_int.Make'.c, Examples.Int_int.Make'.S.t) factory; >