Module Tjr_impfs.V2_usedlist_impl

A persistent list that keeps track of blocks that are allocated to a particular object.

module Usedlist : sig ... end

NOTE the in-memory state of the usedlist is opaque to us; after operations, we check the origin to see if it has changed and then possibly flush/barrier/sync

type ('blk_id, 'blk, 't) usedlist_factory = < with_ : blk_dev_ops:('blk_id, 'blk, 't) Tjr_fs_shared.blk_dev_ops -> barrier:(unit -> (unit, 't) Tjr_monad.m) -> freelist_ops:('blk_id, 'blk_id, 't) Tjr_freelist.freelist_ops -> < usedlist_ops : 'blk_id Usedlist.origin -> (('blk_id, 't) Usedlist.ops, 't) Tjr_monad.m; alloc_via_usedlist : ('blk_id, 't) Usedlist.ops -> ('blk_id, 't) Tjr_fs_shared.blk_allocator_ops; create : unit -> (('blk_id, 't) Usedlist.ops, 't) Tjr_monad.m; >; >

module type S = sig ... end

module type T = sig ... end

module Make_v1 (S : S) : sig ... end

Version with full sig

module Make_v2 (S : S) : T with module S = S

With restricted sig

val usedlist_example : (Tjr_fs_shared.Shared_ctxt.r, Tjr_fs_shared.Shared_ctxt.blk, Tjr_fs_shared.Shared_ctxt.t) usedlist_factory