Module Tjr_impfs__Dir_impl

A simple directory implementation; NOTE this module safe to open.

Directories are one of the main objects we deal with.

We implement a directory as a map from name (string, max length 256 bytes) to entry (file, directory or symlink). Of course, this uses an underlying B-tree.

Any used blocks are recorded in a per-directory-object used list. Currently, blocks are only reclaimed when the directory is deleted. Later, we plan to implement cleaning of directories that are still live, to free up blocks that are no longer used.

We reuse the usedlist implementation from File_impl_v2.

A directory is really a thin layer over a B-tree, interfacing with the used list, and handling the origin block.

type stat_times = Tjr_minifs.Minifs_intf.times

val bin_shape_stat_times : Bin_prot.Shape.t

val bin_size_stat_times : stat_times Bin_prot.Size.sizer

val bin_write_stat_times : stat_times Bin_prot.Write.writer

val bin_writer_stat_times : stat_times Bin_prot.Type_class.writer

val __bin_read_stat_times__ : (int -> stat_times) Bin_prot.Read.reader

val bin_read_stat_times : stat_times Bin_prot.Read.reader

val bin_reader_stat_times : stat_times Bin_prot.Type_class.reader

val bin_stat_times : stat_times Bin_prot.Type_class.t

module Dir_origin : sig ... end

module Dir_ops : sig ... end

module Dir_im : sig ... end

What we keep in memory to represent a dir (in addition to the used list and the dir map)

type ('blk_id, 'blk, 'de, 't, 'did) dir_factory = < read_origin : blk_dev_ops:('blk_id, 'blk, 't) Tjr_fs_shared.blk_dev_ops -> blk_id:'blk_id -> (('blk_id, 'did) Dir_origin.t, 't) Tjr_monad.m; write_origin : blk_dev_ops:('blk_id, 'blk, 't) Tjr_fs_shared.blk_dev_ops -> blk_id:'blk_id -> origin:('blk_id, 'did) Dir_origin.t -> (unit, 't) Tjr_monad.m; with_ : blk_dev_ops:('blk_id, 'blk, 't) Tjr_fs_shared.blk_dev_ops -> barrier:(unit -> (unit, 't) Tjr_monad.m) -> sync:(unit -> (unit, 't) Tjr_monad.m) -> freelist_ops:('blk_id, 'blk_id, 't) Tjr_impfs.Freelist_intf.freelist_ops -> < usedlist_ops : 'blk_id Tjr_impfs.Usedlist_impl.Usedlist.origin -> (('blk_id, 't) Tjr_impfs.Usedlist_impl.Usedlist.ops, 't) Tjr_monad.m; alloc_via_usedlist : ('blk_id, 't) Tjr_impfs.Usedlist_impl.Usedlist.ops -> ('blk_id, 't) Tjr_fs_shared.blk_allocator_ops; mk_dir : usedlist:('blk_id, 't) Tjr_impfs.Usedlist_impl.Usedlist.ops -> btree_root:'blk_id -> with_dir:('did Dir_im.dir_im, 't) Tjr_monad.with_state -> (Tjr_fs_shared.str_256, 'de, 'blk_id, 't, 'did) Dir_ops.t; create_root_dir : root_did:'did -> times:stat_times -> ('blk_id, 't) Tjr_monad.m; create_dir : parent:'did -> times:stat_times -> ('blk_id, 't) Tjr_monad.m; dir_add_autosync : 'blk_id -> (Tjr_fs_shared.str_256, 'de, 'blk_id, 't, 'did) Dir_ops.t -> (Tjr_fs_shared.str_256, 'de, 'blk_id, 't, 'did) Dir_ops.t; dir_from_origin_blk : ('blk_id * ('blk_id, 'did) Dir_origin.t) -> ((Tjr_fs_shared.str_256, 'de, 'blk_id, 't, 'did) Dir_ops.t, 't) Tjr_monad.m; dir_from_origin : 'blk_id -> ((Tjr_fs_shared.str_256, 'de, 'blk_id, 't, 'did) Dir_ops.t, 't) Tjr_monad.m; >; >

NOTE 'de stands for dir_entry

module type S = sig ... end

module type T = sig ... end

module Make_v1 : functor (S : S) -> sig ... end

Make with full sig

module Make_v2 : functor (S : S) -> T with module S = S

Make with restricted sig

module Dir_entry : sig ... end

val dir_example : (Tjr_fs_shared.Shared_ctxt.r, Tjr_fs_shared.ba_buf, Dir_entry.t, Tjr_fs_shared.Shared_ctxt.t, Dir_entry.did) dir_factory