Module Tjr_impfs__File_impl_v2

A simple file implementation; supersedes v1.

Terminology: (file) block-index map: the map from blk index to blk_id

We implement a file as a map from blk index to blk. The map is typically implemented by a B-tree or similar.

New blocks are allocated as needed (by default the map is empty, and a block is only allocated when that block is written to; so files are "sparse").

The "iterated block read/write" component has been abstracted out because it appears in a lot of other places.

Concurrency: The backing B-tree is itself a persistent object.

GOM interaction: We treat a file as a separate object; the GOM maps file id to origin block. For a file impl that uses in-place mutation, we don't need to modify the GOM if the file changes.

TODO:

b-tree to implement delete_after (or some alternative approach)

validate against "normal" file semantics

ensure the tests are run by some top-level test executable

implement freeing of blocks after truncate

module Iter_block_blit = Tjr_impfs.Fv2_iter_block_blit

type idx = Iter_block_blit.idx

val idx_mshlr : Iter_block_blit.Idx.t Tjr_fs_shared.bp_mshlr

type ('buf, 'blk, 'blk_id, 't) file_factory = < read_origin : blk_dev_ops:('blk_id, 'blk, 't) Tjr_fs_shared.blk_dev_ops -> blk_id:'blk_id -> ('blk_id Tjr_impfs.Fv2_types.File_origin_block.t, 't) Tjr_monad.m; write_origin : blk_dev_ops:('blk_id, 'blk, 't) Tjr_fs_shared.blk_dev_ops -> blk_id:'blk_id -> origin:'blk_id Tjr_impfs.Fv2_types.File_origin_block.t -> (unit, 't) Tjr_monad.m; usedlist_origin : 'blk_id Tjr_impfs.Fv2_types.File_origin_block.t -> 'blk_id Tjr_impfs.Fv2_types.Usedlist.origin; with_ : blk_dev_ops:('blk_id, 'blk, 't) Tjr_fs_shared.blk_dev_ops -> barrier:(unit -> (unit, 't) Tjr_monad.m) -> sync:(unit -> (unit, 't) Tjr_monad.m) -> freelist_ops:('blk_id, 't) Tjr_fs_shared.blk_allocator_ops -> < usedlist_ops : 'blk_id Tjr_impfs.Fv2_types.Usedlist.origin -> (('blk_id, 't) Tjr_impfs.Fv2_types.Usedlist.ops, 't) Tjr_monad.m; alloc_via_usedlist : ('blk_id, 't) Tjr_impfs.Fv2_types.Usedlist.ops -> ('blk_id, 't) Tjr_fs_shared.blk_allocator_ops; mk_blk_idx_map : usedlist:('blk_id, 't) Tjr_impfs.Fv2_types.Usedlist.ops -> btree_root:'blk_id -> (int, 'blk_id, 'blk_id, 't) Tjr_impfs.Fv2_types.Btree_ops.t; file_ops : usedlist:('blk_id, 't) Tjr_impfs.Fv2_types.Usedlist.ops -> alloc_via_usedlist:(unit -> ('blk_id, 't) Tjr_monad.m) -> blk_idx_map:(int, 'blk_id, 'blk_id, 't) Tjr_impfs.Fv2_types.Btree_ops.t -> file_origin:'blk_id -> file_size:int -> ('buf, 't) Tjr_impfs.Fv2_types.file_ops; file_from_origin_blk : ('blk_id * 'blk_id Tjr_impfs.Fv2_types.File_origin_block.t) -> (('buf, 't) Tjr_impfs.Fv2_types.file_ops, 't) Tjr_monad.m; file_from_origin : 'blk_id -> (('buf, 't) Tjr_impfs.Fv2_types.file_ops, 't) Tjr_monad.m; >; >

Construct file operations.

Sequence of steps:

(1) Read the file origin block; this contains: file size; the blk-idx map root; the usedlist roots

(2) Construct the usedlist

(2.1) Get the usedlist origin information

(2.2) And this provides the usedlist operations

(2.3) And allocation should indirect via the usedlist (to record which blk_ids have been allocated)

(3) Construct the blk-idx map

(3.1) Construct the ops from the blk-idx root

(4) Finally construct the file operations

val pread_check : size:int Tjr_fs_shared.Int_like.size_ -> off:int Tjr_fs_shared.Int_like.offset_ -> len:'a Tjr_fs_shared.Int_like.len_ -> (unit, string) Stdlib.result

Check the arguments to pread

val pwrite_check : buf_ops:'a Tjr_fs_shared.Buffers_from_btree.buf_ops -> src:'a -> src_off:Tjr_fs_shared.Int_like.offset -> src_len:Tjr_fs_shared.Int_like.len -> dst_off:Tjr_fs_shared.Int_like.offset -> (unit, string) Stdlib.result

Check arguments to pwrite

module type S = sig ... end

module type T = sig ... end

module Make_v1 : functor (S : S) -> sig ... end

Version with full sig