Module Tjr_impfs__File_impl_v2.File_im

State we hold in memory for a particular file

type t = {

file_size : int;

times : stat_times;

last_synced : float;

}

The usedlist and blk-idx map are fixed for the lifetime of the file, so should be parameters on creation. Other than that, we just have a field for file size and a field for the location of the origin blk. The last_synced field is used in conjunction with the last modified field to determine if the file needs syncing.