Module Make_v1.With_

Parameters

S2 : sig ... end

Signature

val blk_sz : int

val usedlist_ops : S.blk_id Tjr_impfs.Fv2_types.Usedlist.origin -> ((S.r, S.t) Tjr_impfs.Fv2_types.Usedlist.ops, S.t) Tjr_monad.m

Implement the usedlist, using the plist and the global freelist. NOTE For the usedlist, we may hold a free block temporarily; if it isn't used we can return it immediately to the global freelist.

val alloc_via_usedlist : (S.blk_id, S.t) Tjr_impfs.Fv2_types.Usedlist.ops -> (S.blk_id, S.t) Tjr_fs_shared.blk_allocator_ops

val mk_blk_idx_map : usedlist:(S.blk_id, S.t) Tjr_impfs.Fv2_types.Usedlist.ops -> btree_root:S.r -> (int, S.r, S.r, S.t) Tjr_impfs.Fv2_types.Btree_ops.t

module File_ops : functor (S3 : sig ... end) -> sig ... end

val file_ops : usedlist:(S.blk_id, S.t) Tjr_impfs.Fv2_types.Usedlist.ops -> alloc_via_usedlist:(unit -> (S.blk_id, S.t) Tjr_monad.m) -> blk_idx_map:(int, S.blk_id, S.blk_id, S.t) Tjr_impfs.Fv2_types.Btree_ops.t -> file_origin:S.blk_id -> file_size:int -> (S.buf, S.t) Tjr_impfs.Fv2_types.file_ops

val file_from_origin_blk : (S.blk_id * S.blk_id Tjr_impfs.Fv2_types.File_origin_block.file_origin_block) -> ((S.buf, S.t) Tjr_impfs.Fv2_types.file_ops, S.t) Tjr_monad.m

val file_from_origin : S.blk_id -> ((S.buf, S.t) Tjr_impfs.Fv2_types.file_ops, S.t) Tjr_monad.m

val export : < alloc_via_usedlist : (S.blk_id, S.t) Tjr_impfs.Fv2_types.Usedlist.ops -> (S.blk_id, S.t) Tjr_fs_shared.blk_allocator_ops; file_from_origin : S.blk_id -> ((S.buf, S.t) Tjr_impfs.Fv2_types.file_ops, S.t) Tjr_monad.m; file_from_origin_blk : (S.blk_id * S.blk_id Tjr_impfs.Fv2_types.File_origin_block.file_origin_block) -> ((S.buf, S.t) Tjr_impfs.Fv2_types.file_ops, S.t) Tjr_monad.m; file_ops : usedlist:(S.blk_id, S.t) Tjr_impfs.Fv2_types.Usedlist.ops -> alloc_via_usedlist:(unit -> (S.blk_id, S.t) Tjr_monad.m) -> blk_idx_map:(int, S.blk_id, S.blk_id, S.t) Tjr_impfs.Fv2_types.Btree_ops.t -> file_origin:S.blk_id -> file_size:int -> (S.buf, S.t) Tjr_impfs.Fv2_types.file_ops; mk_blk_idx_map : usedlist:(S.blk_id, S.t) Tjr_impfs.Fv2_types.Usedlist.ops -> btree_root:S.r -> (int, S.r, S.r, S.t) Tjr_impfs.Fv2_types.Btree_ops.t; usedlist_ops : S.blk_id Tjr_impfs.Fv2_types.Usedlist.origin -> ((S.r, S.t) Tjr_impfs.Fv2_types.Usedlist.ops, S.t) Tjr_monad.m; >