Module Tjr_impfs__Fv2_types

Types for File_impl_v2

type 'fid file_id = {

fid : 'fid;

}

module Usedlist : sig ... end

NOTE the in-memory state of the usedlist is opaque to us; after operations, we check the origin to see if it has changed and then possibly flush/barrier/sync

module Freelist : sig ... end

module Btree_ops : sig ... end

module File_origin_block : sig ... end

We reconstruct the file from the contents of the file's origin blk; this includes the size, the blk->index map root, and the pointers for the used list (which is used to implement alloc in conjunction with the freelist)

module Fo = File_origin_block

module File_im : sig ... end

State we hold in memory for a particular file

type pread_error =

| Pread_error of string

type pwrite_error =

| Pwrite_error of string

type ('buf, 't) file_ops = {

size : unit -> (Tjr_fs_shared.Int_like.size, 't) Tjr_monad.m;

pwrite : src:'buf -> src_off:Tjr_fs_shared.Int_like.offset -> src_len:Tjr_fs_shared.Int_like.len -> dst_off:Tjr_fs_shared.Int_like.offset -> ((Tjr_fs_shared.Int_like.size, pwrite_error) Stdlib.result, 't) Tjr_monad.m;

pread : off:Tjr_fs_shared.Int_like.offset -> len:Tjr_fs_shared.Int_like.len -> (('buf, pread_error) Stdlib.result, 't) Tjr_monad.m;

truncate : size:Tjr_fs_shared.Int_like.size -> (unit, 't) Tjr_monad.m;

flush : unit -> (unit, 't) Tjr_monad.m;

sync : unit -> (unit, 't) Tjr_monad.m;

}

Standard file operations, pwrite, pread, size and truncate.

NOTE we expect buf to be string for the functional version; for mutable buffers we may want to pass the buffer in as a parameter?

NOTE for pwrite, we always return src_len since all bytes are written (unless there is an error of course).

For pread, we always return a buffer of length len.