Module Tjr_impfs__Fv2_types.File_im

State we hold in memory for a particular file

type t = {

file_size : int;

}

The usedlist and blk-idx map are fixed for the lifetime of the file, so should be parameters on creation. Other than that, we just have a single field for file size.