Module Tjr_impfs__Fv2_types.Usedlist

NOTE the in-memory state of the usedlist is opaque to us; after operations, we check the origin to see if it has changed and then possibly flush/barrier/sync

type 'blk_id origin = 'blk_id Tjr_plist.Plist_intf.Pl_origin.pl_origin

val bin_shape_origin : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_origin : blk_id. 'blk_id Bin_prot.Size.sizer -> 'blk_id origin Bin_prot.Size.sizer

val bin_write_origin : blk_id. 'blk_id Bin_prot.Write.writer -> 'blk_id origin Bin_prot.Write.writer

val bin_writer_origin : 'a Bin_prot.Type_class.writer -> 'a origin Bin_prot.Type_class.writer

val __bin_read_origin__ : blk_id. 'blk_id Bin_prot.Read.reader -> (int -> 'blk_id origin) Bin_prot.Read.reader

val bin_read_origin : blk_id. 'blk_id Bin_prot.Read.reader -> 'blk_id origin Bin_prot.Read.reader

val bin_reader_origin : 'a Bin_prot.Type_class.reader -> 'a origin Bin_prot.Type_class.reader

val bin_origin : 'a Bin_prot.Type_class.t -> 'a origin Bin_prot.Type_class.t

type ('blk_id, 't) ops = {

add : 'blk_id -> (unit, 't) Tjr_monad.m;

get_origin : unit -> 'blk_id origin;

flush : unit -> (unit, 't) Tjr_monad.m;

}

The raw operations provided by the plist; in addition we need to integrate the freelist with the usedlist: alloc_via_usedlist

NOTE a sync is just a flush followed by a sync of the underlying blkdev, since we assume all object operations are routed to the same blkdev