Module Tjr_impfs__Fv3

This is a cache-aware implementation of files, for v3.

We model a file as a partial map from blk_off to blk, together with a size in bytes, and file times.

Layers (high to low):

pread,pwrite

aligned block read/write ("pre_file_ops")

cache

raw aligned block read/write

(underlying B-tree implementation or other)

Don't open - too many types with names duplicated elsewhere.

val dont_log : bool

Types

type blk_off = int

module Flag : sig ... end

module Maybe_dirty : sig ... end

type ('k, 'v, 'cache) cache_ops = ('k, 'v, 'cache) Tjr_fs_shared.wbc_ops

type ('blk, 'cache) f_cache = {

times : Tjr_minifs.Minifs_intf.Times.times Maybe_dirty.maybe_dirty;

size : int Maybe_dirty.maybe_dirty;

data : 'cache;

}

Per-file cache entries

type ('blk, 't) lower_ops = {

set_size : int -> (unit, 't) Tjr_monad.m;

set_times : Tjr_minifs.Minifs_intf.times -> (unit, 't) Tjr_monad.m;

write_file_blk : blk_off:int -> 'blk -> (unit, 't) Tjr_monad.m;

read_file_blk : blk_off:int -> ('blk option, 't) Tjr_monad.m;

sync : unit -> (unit, 't) Tjr_monad.m;

}

We need a way to write to the lower non-cached file implementation

type ('blk, 't) pre_file_ops = {

size : unit -> (int, 't) Tjr_monad.m;

truncate : size:int -> (unit, 't) Tjr_monad.m;

get_times : unit -> (Tjr_minifs.Minifs_intf.times, 't) Tjr_monad.m;

set_times : Tjr_minifs.Minifs_intf.times -> (unit, 't) Tjr_monad.m;

write_file_blk : blk_off:int -> 'blk -> (unit, 't) Tjr_monad.m;

read_file_blk : blk_off:int -> ('blk option, 't) Tjr_monad.m;

flush : unit -> (unit, 't) Tjr_monad.m;

sync : unit -> (unit, 't) Tjr_monad.m;

}

type pread_error = Tjr_minifs.Minifs_intf.Call_specific_errors.pread_err

type pwrite_error = Tjr_minifs.Minifs_intf.Call_specific_errors.pwrite_err

type ('buf, 't) file_ops = {

size : unit -> (int, 't) Tjr_monad.m;

truncate : size:int -> (unit, 't) Tjr_monad.m;

get_times : unit -> (Tjr_minifs.Minifs_intf.times, 't) Tjr_monad.m;

set_times : Tjr_minifs.Minifs_intf.times -> (unit, 't) Tjr_monad.m;

pwrite : src:'buf -> src_off:int -> src_len:int -> dst_off:int -> ((int, pwrite_error) Stdlib.result, 't) Tjr_monad.m;

pread : off:int -> len:int -> (('buf, pread_error) Stdlib.result, 't) Tjr_monad.m;

flush : unit -> (unit, 't) Tjr_monad.m;

sync : unit -> (unit, 't) Tjr_monad.m;

}

pread/pwrite argument checks

copied from File_impl_v1

val pread_check : size:int -> off:int -> len:int -> (unit, string) Stdlib.result

Check the arguments to pread

val pwrite_check : buf_ops:'buf Tjr_fs_shared.buf_ops -> src:'buf -> src_off:int -> src_len:int -> dst_off:int -> (unit, string) Stdlib.result

Check arguments to pwrite

Make_v1 - user lower_ops; implement cached pre_file_ops

module Make_v1 : functor (S : sig ... end) -> sig ... end

Make_v2 - use pre_file_ops; implement pread/pwrite using iterated block read/write

module Make_v2 : functor (S : sig ... end) -> sig ... end

Lwt instance

module With_lwt : sig ... end

val file_factory : With_lwt.file_factory