Module Tjr_impfs__Ownership_map

Ownership maps. Each a can own many b; each b is owned by at most one a. If a wants to acquire b, we check b is not already acquired. No locking or anything like that. This simply records the ownership relation. Purely functional.

module Make : functor (S : sig ... end) -> sig ... end