Module Tjr_impfs__V1_resource_manager

This code takes care of the various resources that need to be correctly locked in a concurrent setting.

type tid = {

tid : int;

}

val compare_tid : 'a -> 'a -> int

val new_tid : unit -> (tid, Tjr_monad.With_lwt.lwt) Tjr_monad.m

type live_thread_ops = {

add : tid -> unit;

remove : tid -> unit;

size : unit -> int;

size_below_bound : unit -> bool;

}

val thread_max : int

val live_threads : live_thread_ops

type oid =

| Gom_id

| Fid of int

| Did of int

val compare_oid : 'a -> 'a -> int

type resource_ops = {

owns_nothing : tid -> bool;

acquire : tid:tid -> oid:oid -> (unit, Tjr_fs_shared.Shared_ctxt.t) Tjr_monad.m;

acquire_all : tid:tid -> oids:oid list -> (unit, Tjr_fs_shared.Shared_ctxt.t) Tjr_monad.m;

release : tid:tid -> oid:oid -> (unit, Tjr_fs_shared.Shared_ctxt.t) Tjr_monad.m;

release_all : tid -> (unit, Tjr_fs_shared.Shared_ctxt.t) Tjr_monad.m;

}

acquire and acquire_all require that the thread does not own any resources; this is so that we can guarantee that the resources are obtained in order so we avoid deadlock

val resource_ops : resource_ops

module A : sig ... end

module B : functor (S : sig ... end) -> sig ... end