Module Tjr_impfs__V1_util

module Map_to_set : functor (S : sig ... end) -> sig ... end

module Ownership_map : functor (S : sig ... end) -> sig ... end

Each a can own many b; each b is owned by at most one a. If a wants to acquire b, we check b is not already acquired. No locking or anything like that. This simply records the ownership relation.