Module Tjr_impfs__V2_fs_impl

Implementation of the on-disk filesystem, which is essentially on-disk pointers to the freelist and the GOM

module Fs_origin_block : sig ... end

The origin block for the whole system (typically stored in block 0)

module Pvt : sig ... end

module Filesystem_im : sig ... end

The data we keep in memory, in addition to the data held by the GOM and the freelist. FIXME to avoid repeated syncs, we should append updates to an on-disk log