Module Tjr_impfs__V3_abstract

Abstract development of V3.

See V3.

Initial path resolution is carried out without locks. After, we lock objects that we need to, validate the entries, and then go ahead with the modifications. If we detect changes between path res and execution, we backout with a concurrent modification error.

type times = Tjr_minifs.Minifs_intf.times

module Errors : sig ... end

module Path_resolution : sig ... end

module Lock_ops : sig ... end

module type S0 = sig ... end

module S1 : functor (S0 : S0) -> sig ... end

Types created from basic types

module S2 : functor (S0 : S0) -> sig ... end

The values we expect to be present

module type T = sig ... end

The resulting filesystem (fd is a file identifier, dh is a directory handle supporting leaf_stream operations)

module Make : functor (S0 : S0) -> sig ... end

module X = Make