Module Tjr_impfs__V3_abstract.S1

Types created from basic types

Parameters

S0 : S0

Signature

Directory types

type dir_entry =

| Fid of S0.fid

| Did of S0.did

| Sid of S0.sid

val bin_shape_dir_entry : Bin_prot.Shape.t

val bin_size_dir_entry : dir_entry Bin_prot.Size.sizer

val bin_write_dir_entry : dir_entry Bin_prot.Write.writer

val bin_writer_dir_entry : dir_entry Bin_prot.Type_class.writer

val __bin_read_dir_entry__ : (int -> dir_entry) Bin_prot.Read.reader

val bin_read_dir_entry : dir_entry Bin_prot.Read.reader

val bin_reader_dir_entry : dir_entry Bin_prot.Type_class.reader

val bin_dir_entry : dir_entry Bin_prot.Type_class.t

type dir_k = Tjr_fs_shared.str_256

type dir_v = dir_entry

type dir_ops = {

find : did:S0.did -> Tjr_fs_shared.str_256 -> (dir_entry option, S0.t) Tjr_monad.m;

insert : did:S0.did -> Tjr_fs_shared.str_256 -> dir_entry -> (unit, S0.t) Tjr_monad.m;

delete : did:S0.did -> name:Tjr_fs_shared.str_256 -> obj:dir_entry -> reduce_obj_nlink:unit -> (unit, S0.t) Tjr_monad.m;

set_times : did:S0.did -> times -> (unit, S0.t) Tjr_monad.m;

get_times : did:S0.did -> (times, S0.t) Tjr_monad.m;

set_parent : did:S0.did -> parent:S0.did -> (unit, S0.t) Tjr_monad.m;
NOTE inter-object link

get_parent : did:S0.did -> (S0.did, S0.t) Tjr_monad.m;

sync : did:S0.did -> (unit, S0.t) Tjr_monad.m;
used in rename

}

NOTE the parent field is an inter-object link and so needs to be updated atomically

type rename_case =

| Rename_file_missing of {

times : times;

src : S0.did * Tjr_fs_shared.str_256 * S0.fid;

dst : S0.did * Tjr_fs_shared.str_256;

}
Assumes src and dst are locked and src_name,fid in src; src may equal dst; if so, src_name<>dst_name FIXME check this in impl; tid enable check of locked-ness and contents of src and dst

| Rename_file_file of {

times : times;

src : S0.did * Tjr_fs_shared.str_256 * S0.fid;

dst : S0.did * Tjr_fs_shared.str_256 * S0.fid;

}
fid1<>fid2, so (did1,name1) <> (did2,name2)

| Rename_dir_missing of {

times : times;

src : S0.did * Tjr_fs_shared.str_256 * S0.did;

dst : S0.did * Tjr_fs_shared.str_256;

}
Assumes all locks held, ancestor check completed etc

type dirs_ops = {

delete : S0.did -> (unit, S0.t) Tjr_monad.m;

create : parent:S0.did -> name:Tjr_fs_shared.str_256 -> times:times -> (unit, S0.t) Tjr_monad.m;

rename : rename_case -> (unit, S0.t) Tjr_monad.m;

}

Dir handles

type dhs_ops = {

opendir : S0.did -> (S0.dh, S0.t) Tjr_monad.m;

readdir : S0.dh -> (Tjr_fs_shared.str_256 list, S0.t) Tjr_monad.m;

closedir : S0.dh -> (unit, S0.t) Tjr_monad.m;

}

Types related to files

type fd = S0.fid

type buf = Tjr_fs_shared.ba_buf

type file_ops = {

pread : fid:S0.fid -> foff:int -> len:int -> buf:buf -> boff:int -> ((int, $pread_err) Stdlib.result, S0.t) Tjr_monad.m;

pwrite : fid:S0.fid -> foff:int -> len:int -> buf:buf -> boff:int -> ((int, $pwrite_err) Stdlib.result, S0.t) Tjr_monad.m;

truncate : fid:S0.fid -> int -> (unit, S0.t) Tjr_monad.m;

get_sz : fid:S0.fid -> (int, S0.t) Tjr_monad.m;

set_times : fid:S0.fid -> times -> (unit, S0.t) Tjr_monad.m;

get_times : fid:S0.fid -> (times, S0.t) Tjr_monad.m;

get_nlink : fid:S0.fid -> (int, S0.t) Tjr_monad.m;

set_nlink : fid:S0.fid -> int -> (unit, S0.t) Tjr_monad.m;
NOTE inter-object

sync : fid:S0.fid -> (unit, S0.t) Tjr_monad.m;

}

NOTE the nlink field is derived from inter-object links, and so needs to be updated atomically

type files_ops = {

delete : S0.fid -> (unit, S0.t) Tjr_monad.m;

create : parent:S0.did -> name:Tjr_fs_shared.str_256 -> times:times -> (unit, S0.t) Tjr_monad.m;

create_symlink : parent:S0.did -> name:Tjr_fs_shared.str_256 -> times:times -> contents:Tjr_fs_shared.str_256 -> (unit, S0.t) Tjr_monad.m;

}

NOTE create just creates a new file in the gom; it doesn't link it into a parent etc, so nlink is 0

Path resolution types

type nonrec resolve_path = (S0.fid, S0.did, S0.sid, S0.t) Path_resolution.resolve_path

type path = string

type extra_ops = {

internal_err : a. string -> ('a, S0.t) Tjr_monad.m;

is_ancestor : parent:S0.did -> child:S0.did -> (bool, S0.t) Tjr_monad.m;

}