Module Tjr_impfs_v3.V3_intf

Interfaces describing the abstract development of V3.

Levels/layers are the layers from standard software engineering. For us, level 0 is the topmost, and level 1 sits underneath etc.

Level 0: fuse integration

Level 1: operations, with explicit thread identifiers, used for locking resources

Level 2: operations without thread identifiers; "raw" interface used by level 1

Stages are related to time: stage 1 occurs before stage 2 etc. This recognizes that some values are available only at later points in time.

Preamble

module Errors = Tjr_minifs.Minifs_intf.Error_

val eRROR_CONCURRENT_MODIFICATION : [> `Error_other ]

To avoid adding an error that is not in POSIX, we introduce this pseudo error constructor which just default to `Error_other

NOTE The following are copied from v1_generic.ml

module Path_resolution : sig ... end

module Lock_ops : sig ... end

Basic types, extended types

From a basic set of types, we construct further types.

module type S0 = sig ... end

Very basic types: monadic t, fid, did, sid...

module S1 (S0 : S0) : sig ... end

Types created from basic types

Abstract types describing implementation in terms of levels and stages

module Level2_provides (S0 : S0) : sig ... end

The values we expect to be present at level 2

module Level1_provides : sig ... end

Level 1 provides a single "ops" record of operations, parameterized by thread id tid

module Level0_provides : sig ... end

Level 0 provides a single "ops" record of operations, as Minifs_intf.Ops_type.ops; note that we convert Shared_ctxt.buf to ba_buf