Module V3_intf.S1

Types created from basic types

Directory types

Parameters

module S0 : S0

Signature

type dir_entry =

| Fid of S0.fid

| Did of S0.did

| Sid of S0.sid

val dir_entry_to_yojson : dir_entry -> Yojson.Safe.t

val dir_entry_of_yojson : Yojson.Safe.t -> dir_entry Ppx_deriving_yojson_runtime.error_or

val _ : Yojson.Safe.t -> dir_entry Ppx_deriving_yojson_runtime.error_or

val bin_shape_dir_entry : Bin_prot.Shape.t

val bin_size_dir_entry : dir_entry Bin_prot.Size.sizer

val bin_write_dir_entry : dir_entry Bin_prot.Write.writer

val bin_writer_dir_entry : dir_entry Bin_prot.Type_class.writer

val __bin_read_dir_entry__ : (int -> dir_entry) Bin_prot.Read.reader

val bin_read_dir_entry : dir_entry Bin_prot.Read.reader

val bin_reader_dir_entry : dir_entry Bin_prot.Type_class.reader

val bin_dir_entry : dir_entry Bin_prot.Type_class.t

type dir_k = Tjr_fs_shared.str_256

type dir_v = dir_entry

type dir_ops = {

find : did:S0.did -> Tjr_fs_shared.str_256 -> (dir_entry option, S0.t) Tjr_monad.m;

insert : did:S0.did -> Tjr_fs_shared.str_256 -> dir_entry -> (unit, S0.t) Tjr_monad.m;

delete : did_locked:unit -> did:S0.did -> name:Tjr_fs_shared.str_256 -> obj:dir_entry -> reduce_obj_nlink:unit -> (unit, S0.t) Tjr_monad.m;

set_times : did:S0.did -> Tjr_fs_shared.Times.times -> (unit, S0.t) Tjr_monad.m;

get_times : did:S0.did -> (Tjr_fs_shared.Times.times, S0.t) Tjr_monad.m;

get_parent : did:S0.did -> (S0.did, S0.t) Tjr_monad.m;

sync : lock_held:bool -> did:S0.did -> (unit, S0.t) Tjr_monad.m;	(*
used in rename
*)

}

Operations on a single directory

NOTE the parent field is an inter-object link and so needs to be updated atomically ; NOTE no nlink for dirs, but POSIX probably requires it?

type rename_file =

| Rename_file_missing of {

locks_held : unit;

times : Tjr_fs_shared.Times.times;

src : S0.did * Tjr_fs_shared.str_256 * S0.fid;

dst : S0.did * Tjr_fs_shared.str_256;

}	(*
Assumes src and dst are locked and src_name,fid in src; src may equal dst; if so, src_name<>dst_name FIXME check this in impl; tid enable check of locked-ness and contents of src and dst
*)

| Rename_file_file of {

locks_held : unit;

times : Tjr_fs_shared.Times.times;

src : S0.did * Tjr_fs_shared.str_256 * S0.fid;

dst : S0.did * Tjr_fs_shared.str_256 * S0.fid;

}	(*
fid1<>fid2, so (did1,name1) <> (did2,name2)
*)

type rename_dir_missing =

| Rename_dir_missing of {

tid : S0.tid;

locks_not_held : unit;

needs_ancestor_check : unit;

src : S0.did * Tjr_fs_shared.str_256 * S0.did;

dst : S0.did * Tjr_fs_shared.str_256;

}	(*
Implementation should incorporate ancestor check,locking etc etc
*)

type dirs_ops = {

delete : S0.did -> (unit, S0.t) Tjr_monad.m;

create_and_add_to_parent : parent_locked:unit -> parent:S0.did -> name:Tjr_fs_shared.str_256 -> times:Tjr_fs_shared.Times.times -> (unit, S0.t) Tjr_monad.m;

rename_file : rename_file -> (unit, S0.t) Tjr_monad.m;

rename_dir : rename_dir_missing -> ((unit, Errors.exn_) Stdlib.result, S0.t) Tjr_monad.m;

}

Operations to create/delete a directory, and operate on more than one directory (eg cross-directory rename)

Dir handles

type dir_handle_ops = {

opendir : tid:S0.tid -> S0.did -> (S0.dh, S0.t) Tjr_monad.m;

readdir : tid:S0.tid -> S0.dh -> (Tjr_fs_shared.str_256 list, S0.t) Tjr_monad.m;

closedir : tid:S0.tid -> S0.dh -> (unit, S0.t) Tjr_monad.m;

}

Types related to files

type fd = S0.fid

type buf = Tjr_fs_shared.Shared_ctxt.buf

type file_ops = {

pread : fid:S0.fid -> foff:int -> len:int -> buf:buf -> boff:int -> ((int, Tjr_minifs.Minifs_intf.Call_specific_errors.pread_err) Stdlib.result, S0.t) Tjr_monad.m;

pwrite : fid:S0.fid -> foff:int -> len:int -> buf:buf -> boff:int -> ((int, Tjr_minifs.Minifs_intf.Call_specific_errors.pwrite_err) Stdlib.result, S0.t) Tjr_monad.m;

truncate : fid:S0.fid -> int -> (unit, S0.t) Tjr_monad.m;

get_sz : fid:S0.fid -> (int, S0.t) Tjr_monad.m;

set_times : fid:S0.fid -> Tjr_fs_shared.Times.times -> (unit, S0.t) Tjr_monad.m;

get_times : fid:S0.fid -> (Tjr_fs_shared.Times.times, S0.t) Tjr_monad.m;

sync : fid:S0.fid -> (unit, S0.t) Tjr_monad.m;

}

NOTE the nlink field is derived from inter-object links, and so needs to be updated atomically

type files_ops = {

create_and_add_to_parent : parent:S0.did -> name:Tjr_fs_shared.str_256 -> times:Tjr_fs_shared.Times.times -> (unit, S0.t) Tjr_monad.m;

create_symlink_and_add_to_parent : parent:S0.did -> name:Tjr_fs_shared.str_256 -> times:Tjr_fs_shared.Times.times -> contents:Tjr_fs_shared.str_256 -> (unit, S0.t) Tjr_monad.m;

}

Path resolution types

type nonrec resolve_path = (S0.fid, S0.did, S0.sid, S0.t) Path_resolution.resolve_path

type path = string

type extra_ops = {

internal_err : a. tid:S0.tid -> string -> ('a, S0.t) Tjr_monad.m;

}