Module Tjr_impfs_v3.V3_level1

Level 1: build on level 2 to provide essential filesystem functionality.

NOTE per-thread locking is provided at level 1 (since level 2 has no notion of tid... except for operations like Rename_dir_missing where a tid is provided.).

Level 1: Initially path resolution is carried out without locks. Afterwards, we lock objects that we need to, validate the entries, and then go ahead with the modifications. If we detect changes between path res and execution, we backout with a concurrent modification error.

val line : int -> bool

module Make (S0 : V3_intf.S0) : sig ... end