Module Tjr_impfs_v3.V3_level2

V3 level 2 is the lowest implementation level; SQLite is used to implement directories and metadata (?) and some other filesystem is used to implement files.

This version focuses on the front end caching and locking, so we delegate dirs and files to SQLite/the existing filesystem.

We maintain a cache of live files, live dirs.

We use kref reference counting to ensure that objects are not removed from the cache whilst we operate on them. Locked objects can never be removed from the cache, but we maintain an invariant that any locked object must have an existing kref, so it should never be the case that there are no references but an object is locked.

In order to operate on an object, we first obtain a kref: krefs.get id. This ensures that the object is in the cache and will not be flushed.

Preamble

val line : int -> bool

module R : sig ... end

module S0 = V3_base_types

module S1 : sig ... end

module Level2_provides : sig ... end

module type T2 = Level2_provides.T2

What we have to implement - see Stage2 below.

Directories

type config = {

entries_cache_capacity : int;

entries_cache_trim_delta : int;

live_dirs_capacity : int;

live_dirs_trim_delta : int;

file_data_path : string;

live_files_capacity : int;

live_files_trim_delta : int;

}

Configuration params known at module init time

module Directories_prelude : sig ... end

NOTE nothing in here is actually exposed in Stage2 below; could just open the structure rather than binding it

module type STAGE1 = sig ... end

Stage 1 is immediately after connection to the SQLite backend; it includes the sql_dir_ops live connection to the database

module Stage2 (Stage1 : STAGE1) : T2

Stage 2 uses the connection to SQLite to build the rest of the T2 interface