Module V3_level2.Directories_prelude

NOTE nothing in here is actually exposed in Stage2 below; could just open the structure rather than binding it

module Lru : sig ... end

val lru : (Tjr_fs_shared.str_256, S1.dir_entry, unit) Lru.lru_module

module Entries_cache : sig ... end

type entries_cache = Entries_cache.cache_state'

val entries_cache_ops : (Entries_cache.k, Entries_cache.v, Entries_cache.cache_state', Entries_cache.bot, Entries_cache.t) Lru.Top.ops

type sql_dir_ops = (Tjr_fs_shared.str_256, S1.dir_entry, S0.t, S0.did) V3_sqlite.Sqlite_dir.dir_ops

type per_dir = {

lock : Tjr_monad.With_lwt.lwt_mutex;

parent : S0.did R.ref;

times : Tjr_minifs.Minifs_intf.times R.ref;

entries_cache : entries_cache;

}

Live dirs; use V3_live_object_cache

module S2 : sig ... end

type per_file = {

fd : Lwt_unix.file_descr;

}