Module Lru.Bot

Lowest level (bottom) operations, typically those supported by a persistent on-disk log.

type ('k, 'v, 'bot, 't) ops = {

find_opt : 'bot -> 'k -> ('v option, 't) Tjr_monad.m;

sync : 'bot -> (unit, 't) Tjr_monad.m;

exec : 'bot -> sync:bool -> ('k, 'v) op list -> (unit, 't) Tjr_monad.m;

}

We hope that the operations are executed linearly; it should be the case that operations (apart from find_opt) on the bottom level always need the "is_syncing" lock held. In this case, exec should presumably always be sync? Or do we allow some buffering here as well, in the case that sync flag is not set? And does an exec force other outstanding operations to sync, or just those in the list given? Or is the semantics that sync is called immediately afterwards, so everything is synced? Yes, this semantics seems simpler. FIXME rename exec to exec_sync and remove sync flag