Module Lru.Top

Top-level operations.

type ('k, 'v, 'c, 'bot, 't) ops = {

initial_cache_state : cap:int -> trim_delta:int -> bot:'bot -> bot_ops:('k, 'v, 'bot, 't) Bot.ops -> 'c;

find_opt : 'c -> 'k -> ('v option, 't) Tjr_monad.m;

insert : 'c -> 'k -> 'v -> (unit, 't) Tjr_monad.m;

delete : 'c -> 'k -> (unit, 't) Tjr_monad.m;

unsafe_clean : 'c -> ('k, 'v) op list;	(*
unsafe_clean removes all entries and returns them as a list; this assumes the caller has control over concurrency (ie that no-one is using the cache, and the cache is not syncing), and will sync to the lower layer themselves; it is used to avoid multiple syncs to the lower level when syncing eg a dir's dirty entries and a dir's meta, see v3_level2.ml
*)

sync : 'c -> (unit, 't) Tjr_monad.m;

exec : 'c -> sync:bool -> ('k, 'v) op list -> (unit, 't) Tjr_monad.m;

remove_clean_entry : 'c -> 'k -> unit;	(*
NOTE this assumes that the cache is locked, not syncing, and no other operations are pending
*)

}