Module V3_level2.Lru

Interfaces

type ('k, 'v) op = [

| `Insert of 'k * 'v

| `Delete of 'k

]

type 'v op' = [

| `Insert of 'v

| `Delete

]

module Bot : sig ... end

Lowest level (bottom) operations, typically those supported by a persistent on-disk log.

module Top : sig ... end

Top-level operations.

Implementation

type 'v entry' = 'v option * bool Stdlib.ref

type ('v, 't) entry =

| Plain of 'v entry'	(*
the ref is "dirty" flag
*)

| Finding of {

time : int;

promise : ('v option, 't) Tjr_monad.m;

}	(*
We use the time to distinguish between concurrent finds that return out of order (FIXME no longer used)
*)

val lower_none : unit -> 'a option * bool Stdlib.ref

val lower_some : 'a -> 'a option * bool Stdlib.ref

val insert : 'a -> 'a option * bool Stdlib.ref

val delete : unit -> 'a option * bool Stdlib.ref

type ('k, 'v, 'lru, 'bot, 't) cache_state = {

bot : 'bot;

bot_ops : ('k, 'v, 'bot, 't) Bot.ops;

mutable is_syncing : bool;

mutable syncing : (unit, 't) Tjr_monad.m;	(*
If we are syncing, then we service operations we can immediately, and other operations wait for the sync. This field is only meaningful if is_syncing is true.
*)

lru : 'lru;	(*
Lru state is mutable
*)

trim_delta : int;	(*
Number of entries we trim each time the cache gets too big; trim_delta should be much less than cap
*)

lru_capacity : int;	(*
LRU capacity
*)

}

module type S = sig ... end

module type T = sig ... end

module Make (S : S) : T with type k = S.k and type v = S.v and type lru = S.lru and type bot = S.bot

Functor make

type ('k, 'v, 'bot) lru_module = (module T with type bot = 'bot and type k = 'k and type v = 'v)

val make : unit -> (module T with type bot = 'bot and type k = 'k and type v = 'v)

make as a function.

NOTE: This uses Stdlib pervasive equality and hashing on values of type k; this provides a simpler interface where the Lru functionality is constructed for the user

module Test () : sig ... end