Module Tjr_impfs_v3.V3_live_object_cache

Live object cache.

A cache of "live" objects. Expunging an object may take some time. Resurrecting an object may take some time. We use reference counting to prevent objects from being expunged while references remain live.

We are maintaining a cache of objects by id. We can maintain the counter with the object. A kref is an abstract wrapper round the object. Essentially it provides: kref_to_obj: kref -> obj. It also includes put: kref -> unit which destroys the kref and reduces the underlying count.

Don't open.

val line : int -> bool

type ('obj, 'kref) kref_ops = {

kref_to_obj : 'kref -> 'obj;

krelease : 'kref -> unit;

}

User view of krefs; the user must call krelease when finished with the object.

module Private_kref_impl : sig ... end

type 'a kref = 'a Private_kref_impl.kref

val kref_ops : ('a, 'a Private_kref_impl.kref) kref_ops

module Kref = Private_kref_impl

A cache of live objects is just an LRU of ref counted objects

type 'a counted = {

count : int Stdlib.ref;

obj : 'a;

}

type ('a, 't) entry = [

| `Resurrecting of (unit, 't) Tjr_monad.m

| `Present of 'a counted

| `Finalising of (unit, 't) Tjr_monad.m

]

type config = {

cache_size : int;

trim_delta : int;

}

type ('lru, 't) cache = {

lru : 'lru;

config : config;

mutable gc_thread : (unit, 't) Tjr_monad.m;

}

type ('id, 'a, 'kref, 'cache, 't) cache_ops = {

create : config:config -> 'cache;

get : 'id -> 'cache -> ('kref, 't) Tjr_monad.m;

put : 'kref -> unit;

kref_to_obj : 'kref -> 'a;

}

module type S = sig ... end

module type T = sig ... end

module Make (S : S) : T with type id = S.id and type a = S.a