Module type V3_live_object_cache.T

type t = Tjr_monad.lwt

type id

type a

type kref

type cache

val make_cache_ops : resurrect:(id -> (a, t) Tjr_monad.m) -> finalise:((id * a) list -> (unit, t) Tjr_monad.m) -> (id, a, kref, cache, t) cache_ops

resurrect: slow operation to resurrect an object from a persistent store; finalise:called when removing objects from the cache; no outstanding references remain; as such, the object should certainly not be locked