Module V3_sqlite.Sqlite_dir

An implementation of directories on top of SQLite, for v3.

Directories consist of two parts:

The metadata: times (atim,mtim), parent

The contents: entries pointing to other files and directories

To implement directories, we need to implement the interface Dir_ops.

type ('k,'v,'t,'did) dir_ops = {
  get_meta  : did:'did -> ('did * times,'t)m;
  find      : did:'did -> 'k -> ('v option,'t) m;
  exec      : ('k,'v,'did) op list -> (unit,'t)m;
  opendir   : did:'did -> (('k,'v,'t)Ls.ops,'t)m;
  max_did   : unit -> 'did; 
  (** NOTE this is blocking, but is intended to be used once at
      startup, so potentially only results in slightly slower startup
      than would be possible if this was in the monad *)

  pre_create: 
    note_does_not_touch_parent: unit -> 
    new_did:'did -> parent:'did -> times:times -> (unit,'t)m; 
  (** pre-create does not touch the parent directory - that needs to
      be done with a separate insert *)
}

We implement this without a cache: all operations go directly to the database.

Let's assume that each directory has a unique (int) identifier did. Then we can have a single table dir_entries, with rows (did, name, entry) indexed by did,name (name is unique of course, for a particular did).

Then find, insert and delete are fine.

For parent and times, we can have another table (did,parent,times).

opendir creates a reference which pulls upto some limit of elts from the db, and stores the max name so we can resume later on ls_step (the directory entries are stored ordered by name in the database).

NOTE currently all DB access is serialized; but some concurrency could be achieved

module Times = Tjr_fs_shared.Times

type times = Times.times

module Ls : sig ... end

Directory listing (like a dir handle)

module Op : sig ... end

Directory operations, to be executed as part of a batch

module Dir_ops : sig ... end

include module type of struct include Dir_ops end

type ('k, 'v, 't, 'did) dir_ops = ('k, 'v, 't, 'did) Dir_ops.dir_ops = {

get_meta : did:'did -> ('did * times, 't) Tjr_monad.m;

find : did:'did -> 'k -> ('v option, 't) Tjr_monad.m;

exec : ('k, 'v, 'did) Op.op list -> (unit, 't) Tjr_monad.m;

opendir : did:'did -> (('k, 'v, 't) Ls.ops, 't) Tjr_monad.m;

max_did : unit -> 'did;

pre_create : note_does_not_touch_parent:unit -> new_did:'did -> parent:'did -> times:times -> (unit, 't) Tjr_monad.m;

}

NOTE 'k is typically "name" of directory entry (a short string); 'v is the directory entry itself (a pointer to a file/dir/symlink); we don't assume much about a dir_entry apart from that we can convert it to/from a string/blob.

Fields are:

get_meta - parent and time info

find a particular entry

exec a list of operations (insert, delete, set parent, set times)

opendir - return a list of entries

max_did - used once on startup, so we can allocate fresh dids; NOTE this is blocking, but is intended to be used once at startup, so potentially only results in slightly slower startup than would be possible if this was in the monad

pre_create which creates a new directory, with a given parent, but does not touch the parent (that has to be done separately)

type db = Sqlite3.db

val create_tables : Sqlite3.db -> unit

val add_root_directory : Sqlite3.db -> unit

module type S = sig ... end

What the following Make functor requires; essentially a way to convert dir_entry to/from string.

module type T = sig ... end

What the following Make functor provides

module Make (S : S) : T with type did := S.did and type dir_entry := S.dir_entry

module Test () : sig ... end