Module Sqlite_dir.Dir_ops

type ('k, 'v, 't, 'did) dir_ops = {

get_meta : did:'did -> ('did * times, 't) Tjr_monad.m;

find : did:'did -> 'k -> ('v option, 't) Tjr_monad.m;

exec : ('k, 'v, 'did) Op.op list -> (unit, 't) Tjr_monad.m;

opendir : did:'did -> (('k, 'v, 't) Ls.ops, 't) Tjr_monad.m;

max_did : unit -> 'did;

pre_create : note_does_not_touch_parent:unit -> new_did:'did -> parent:'did -> times:times -> (unit, 't) Tjr_monad.m;

}

NOTE 'k is typically "name" of directory entry (a short string); 'v is the directory entry itself (a pointer to a file/dir/symlink); we don't assume much about a dir_entry apart from that we can convert it to/from a string/blob.

Fields are:

get_meta - parent and time info

find a particular entry

exec a list of operations (insert, delete, set parent, set times)

opendir - return a list of entries

max_did - used once on startup, so we can allocate fresh dids; NOTE this is blocking, but is intended to be used once at startup, so potentially only results in slightly slower startup than would be possible if this was in the monad

pre_create which creates a new directory, with a given parent, but does not touch the parent (that has to be done separately)