Module Tjr_impfs.File_impl_v2

File types

module Usedlist = Usedlist_impl.Usedlist

module Times = Tjr_minifs.Minifs_intf.Times

type stat_times = Times.times

val bin_shape_stat_times : Bin_prot.Shape.t

val bin_size_stat_times : stat_times Bin_prot.Size.sizer

val bin_write_stat_times : stat_times Bin_prot.Write.writer

val bin_writer_stat_times : stat_times Bin_prot.Type_class.writer

val __bin_read_stat_times__ : (int -> stat_times) Bin_prot.Read.reader

val bin_read_stat_times : stat_times Bin_prot.Read.reader

val bin_reader_stat_times : stat_times Bin_prot.Type_class.reader

val bin_stat_times : stat_times Bin_prot.Type_class.t

val now : unit -> float

val update_atim : Times.times -> Times.times

val update_mtim : Times.times -> Times.times

val new_times : unit -> Times.times

module File_origin_block : sig ... end

We reconstruct the file from the contents of the file's origin blk; this includes the size, the blk->index map root, and the pointers for the used list (which is used to implement alloc in conjunction with the freelist)

module Fo = File_origin_block

module File_im : sig ... end

State we hold in memory for a particular file

type pread_error = Tjr_minifs.Minifs_intf.Call_specific_errors.pread_err

type pwrite_error = Tjr_minifs.Minifs_intf.Call_specific_errors.pwrite_err

type ('blk_id, 'buf, 't) file_ops = {

size : unit -> (int, 't) Tjr_monad.m;

pwrite : src:'buf -> src_off:Tjr_fs_shared.Int_like.offset -> src_len:Tjr_fs_shared.Int_like.len -> dst_off:Tjr_fs_shared.Int_like.offset -> ((int, pwrite_error) Stdlib.result, 't) Tjr_monad.m;

pread : off:Tjr_fs_shared.Int_like.offset -> len:Tjr_fs_shared.Int_like.len -> (('buf, pread_error) Stdlib.result, 't) Tjr_monad.m;

truncate : size:int -> (unit, 't) Tjr_monad.m;

flush : unit -> (unit, 't) Tjr_monad.m;

sync : unit -> (unit, 't) Tjr_monad.m;

get_origin : unit -> ('blk_id File_origin_block.t, 't) Tjr_monad.m;

get_times : unit -> (Tjr_minifs.Minifs_intf.times, 't) Tjr_monad.m;

set_times : Tjr_minifs.Minifs_intf.times -> (unit, 't) Tjr_monad.m;

}

Standard file operations, pwrite, pread, size and truncate.

NOTE we expect buf to be string for the functional version; for mutable buffers we may want to pass the buffer in as a parameter?

NOTE for pwrite, we always return src_len since all bytes are written (unless there is an error of course).

For pread, we always return a buffer of length len (assuming off+len <= file size).

NOTE origin/flush/sync behaviour: flush forces changes (including origin) to disk and issues a barrier; sync ditto, and issues a sync; otherwise, changes that affect the origin should be accompanied by a barrier -> origin-write -> barrier posthook FIXME are they?

module Iter_block_blit = Fv2_iter_block_blit

type idx = Iter_block_blit.idx

val idx_mshlr : Iter_block_blit.Idx.t Tjr_fs_shared.bp_mshlr

type ('buf, 'blk, 'blk_id, 't) file_factory = < read_origin : blk_dev_ops:('blk_id, 'blk, 't) Tjr_fs_shared.blk_dev_ops -> blk_id:'blk_id -> ('blk_id File_origin_block.t, 't) Tjr_monad.m; write_origin : blk_dev_ops:('blk_id, 'blk, 't) Tjr_fs_shared.blk_dev_ops -> blk_id:'blk_id -> origin:'blk_id File_origin_block.t -> (unit, 't) Tjr_monad.m; usedlist_origin : 'blk_id File_origin_block.t -> 'blk_id Fv2_types.Usedlist.origin; with_ : blk_dev_ops:('blk_id, 'blk, 't) Tjr_fs_shared.blk_dev_ops -> barrier:(unit -> (unit, 't) Tjr_monad.m) -> sync:(unit -> (unit, 't) Tjr_monad.m) -> freelist_ops:('blk_id, 'blk_id, 't) Tjr_plist_freelist.Freelist_intf.freelist_ops -> < usedlist_ops : 'blk_id Fv2_types.Usedlist.origin -> (('blk_id, 't) Fv2_types.Usedlist.ops, 't) Tjr_monad.m; alloc_via_usedlist : ('blk_id, 't) Fv2_types.Usedlist.ops -> ('blk_id, 't) Tjr_fs_shared.blk_allocator_ops; mk_blk_idx_map : usedlist:('blk_id, 't) Fv2_types.Usedlist.ops -> btree_root:'blk_id -> (int, 'blk_id, 'blk_id, 't) Fv2_types.Btree_ops.t; file_ops : usedlist:('blk_id, 't) Fv2_types.Usedlist.ops -> alloc_via_usedlist:(unit -> ('blk_id, 't) Tjr_monad.m) -> blk_idx_map:(int, 'blk_id, 'blk_id, 't) Fv2_types.Btree_ops.t -> file_origin:'blk_id -> init_file_size:int -> init_times:Fv2_types.stat_times -> ('blk_id, 'buf, 't) file_ops; create_file : Fv2_types.stat_times -> ('blk_id, 't) Tjr_monad.m; file_from_origin_blk : ('blk_id * 'blk_id File_origin_block.t) -> (('blk_id, 'buf, 't) file_ops, 't) Tjr_monad.m; file_from_origin : 'blk_id -> (('blk_id, 'buf, 't) file_ops, 't) Tjr_monad.m; >; >

Construct file operations.

Sequence of steps:

(1) Read the file origin block; this contains: file size; the blk-idx map root; the usedlist roots

(2) Construct the usedlist

(2.1) Get the usedlist origin information

(2.2) And this provides the usedlist operations

(2.3) And allocation should indirect via the usedlist (to record which blk_ids have been allocated)

(3) Construct the blk-idx map

(3.1) Construct the ops from the blk-idx root

(4) Finally construct the file operations

val pread_check : size:int Tjr_fs_shared.Int_like.size_ -> off:int Tjr_fs_shared.Int_like.offset_ -> len:'a Tjr_fs_shared.Int_like.len_ -> (unit, string) Stdlib.result

Check the arguments to pread

val pwrite_check : buf_ops:'a Tjr_fs_shared.Buffers_from_btree.buf_ops -> src:'a -> src_off:Tjr_fs_shared.Int_like.offset -> src_len:Tjr_fs_shared.Int_like.len -> dst_off:Tjr_fs_shared.Int_like.offset -> (unit, string) Stdlib.result

Check arguments to pwrite

module type S = sig ... end

module type T = sig ... end

module Make_v1 : functor (S : S) -> sig ... end

Version with full sig; NOTE that this is quite specialized to Shared_ctxt

module Make_v2 : functor (S : S) -> T with module S = S

Version with restricted sig

val file_examples : < example_1 : (Tjr_fs_shared.ba_buf, Tjr_fs_shared.ba_buf, Tjr_fs_shared.Shared_ctxt.r, Tjr_fs_shared.Shared_ctxt.t) file_factory; example_2 : (Tjr_fs_shared.ba_buf, Tjr_fs_shared.ba_buf, Tjr_fs_shared.Shared_ctxt.r, Tjr_fs_shared.Shared_ctxt.t) file_factory; >

example_1: (ba_buf, ba_buf, Shared_ctxt.r, Shared_ctxt.t) file_factory (blk_sz 4096)

example_2: (ba_buf, ba_buf, Shared_ctxt.r, Shared_ctxt.t) file_factory (blk_sz 2, for testing)

module Test : functor () sig ... end

Test the file_example with