Module Fl_example_1.With_

Parameters

S : sig ... end

Signature

val pl_factory : Tjr_fs_shared.Shared_ctxt.r Tjr_plist__.Pl_type_abbrevs.plist_factory

val pl_with_ : < add_origin : < set_and_sync : Tjr_plist__.Pl_type_abbrevs.blk_id Tjr_plist__.Plist_intf.Pl_origin.pl_origin -> (unit, Tjr_plist__.Pl_type_abbrevs.t) Tjr_monad.m; > -> (Tjr_fs_shared.Shared_ctxt.r, Tjr_fs_shared.ba_buf, Tjr_plist__.Pl_type_abbrevs.blk_id, Tjr_plist__.Pl_type_abbrevs.t) Tjr_plist__.Plist_intf.plist_ops -> (Tjr_fs_shared.Shared_ctxt.r, Tjr_fs_shared.ba_buf, Tjr_plist__.Pl_type_abbrevs.blk_id, Tjr_plist__.Pl_type_abbrevs.t) Tjr_plist__.Plist_intf.plist_ops; create : Tjr_plist__.Pl_type_abbrevs.blk_id -> (< plist_ops : (Tjr_fs_shared.Shared_ctxt.r, Tjr_fs_shared.ba_buf, Tjr_plist__.Pl_type_abbrevs.blk_id, Tjr_plist__.Pl_type_abbrevs.t) Tjr_plist__.Plist_intf.plist_ops; plist_ref : (Tjr_plist__.Pl_type_abbrevs.blk_id, Tjr_fs_shared.ba_buf) Tjr_plist__.Plist_intf.plist Stdlib.ref; with_plist : ((Tjr_plist__.Pl_type_abbrevs.blk_id, Tjr_fs_shared.ba_buf) Tjr_plist__.Plist_intf.plist, Tjr_plist__.Pl_type_abbrevs.t) Tjr_monad.with_state; >, Tjr_plist__.Pl_type_abbrevs.t) Tjr_monad.m; init : < create : Tjr_plist__.Pl_type_abbrevs.blk_id -> ((Tjr_plist__.Pl_type_abbrevs.blk_id, Tjr_fs_shared.ba_buf) Tjr_plist__.Plist_intf.plist, Tjr_plist__.Pl_type_abbrevs.t) Tjr_monad.m; from_endpts : Tjr_plist__.Pl_type_abbrevs.blk_id Tjr_plist__.Plist_intf.Pl_origin.pl_origin -> ((Tjr_plist__.Pl_type_abbrevs.blk_id, Tjr_fs_shared.ba_buf) Tjr_plist__.Plist_intf.plist, Tjr_plist__.Pl_type_abbrevs.t) Tjr_monad.m; read_from_hd : Tjr_plist__.Pl_type_abbrevs.blk_id -> ((Tjr_fs_shared.Shared_ctxt.r list * Tjr_plist__.Pl_type_abbrevs.blk_id option) list, Tjr_plist__.Pl_type_abbrevs.t) Tjr_monad.m; >; restore : Tjr_plist__.Pl_type_abbrevs.blk_id Tjr_plist__.Plist_intf.Pl_origin.t -> (< plist_ops : (Tjr_fs_shared.Shared_ctxt.r, Tjr_fs_shared.ba_buf, Tjr_plist__.Pl_type_abbrevs.blk_id, Tjr_plist__.Pl_type_abbrevs.t) Tjr_plist__.Plist_intf.plist_ops; plist_ref : (Tjr_plist__.Pl_type_abbrevs.blk_id, Tjr_fs_shared.ba_buf) Tjr_plist__.Plist_intf.plist Stdlib.ref; with_plist : ((Tjr_plist__.Pl_type_abbrevs.blk_id, Tjr_fs_shared.ba_buf) Tjr_plist__.Plist_intf.plist, Tjr_plist__.Pl_type_abbrevs.t) Tjr_monad.with_state; >, Tjr_plist__.Pl_type_abbrevs.t) Tjr_monad.m; with_ref : (Tjr_plist__.Pl_type_abbrevs.blk_id, Tjr_fs_shared.ba_buf) Tjr_plist__.Plist_intf.plist -> < plist_ref : (Tjr_plist__.Pl_type_abbrevs.blk_id, Tjr_fs_shared.ba_buf) Tjr_plist__.Plist_intf.plist Stdlib.ref; with_plist : ((Tjr_plist__.Pl_type_abbrevs.blk_id, Tjr_fs_shared.ba_buf) Tjr_plist__.Plist_intf.plist, Tjr_plist__.Pl_type_abbrevs.t) Tjr_monad.with_state; >; with_state : ((Tjr_plist__.Pl_type_abbrevs.blk_id, Tjr_fs_shared.ba_buf) Tjr_plist__.Plist_intf.plist, Tjr_plist__.Pl_type_abbrevs.t) Tjr_monad.with_state -> (Tjr_fs_shared.Shared_ctxt.r, Tjr_fs_shared.ba_buf, Tjr_plist__.Pl_type_abbrevs.blk_id, Tjr_plist__.Pl_type_abbrevs.t) Tjr_plist__.Plist_intf.plist_ops; >

val plist_ops : Tjr_plist__.Pl_type_abbrevs.blk_id Tjr_plist.Plist_intf.Pl_origin.pl_origin -> ((Tjr_fs_shared.Shared_ctxt.r, Tjr_fs_shared.ba_buf, Tjr_plist__.Pl_type_abbrevs.blk_id, Tjr_plist__.Pl_type_abbrevs.t) Tjr_plist__.Plist_intf.plist_ops, Tjr_monad.With_lwt.lwt) Tjr_monad.m

val min_free_alloc : Tjr_fs_shared.Shared_ctxt.r -> int -> Tjr_fs_shared.Shared_ctxt.B.blk_id Tjr_lib.List_.t * Tjr_fs_shared.Shared_ctxt.B.blk_id

val with_plist_ops : (Tjr_fs_shared.Shared_ctxt.r, 'a, Tjr_fs_shared.Shared_ctxt.r, Tjr_monad.With_lwt.lwt) Tjr_plist.Plist_intf.plist_ops -> < with_locked_ref : Tjr_fs_shared.Shared_ctxt.r Freelist_intf.freelist_im -> < freelist_ops : (Tjr_fs_shared.Shared_ctxt.r, Tjr_fs_shared.Shared_ctxt.r, Tjr_monad.With_lwt.lwt) Freelist_intf.freelist_ops; freelist_ref : Tjr_fs_shared.Shared_ctxt.r Freelist_intf.freelist_im Stdlib.ref; >; with_state : (Tjr_fs_shared.Shared_ctxt.r Freelist_intf.freelist_im, Tjr_monad.With_lwt.lwt) Tjr_monad.with_state -> (Tjr_fs_shared.Shared_ctxt.r, Tjr_fs_shared.Shared_ctxt.r, Tjr_monad.With_lwt.lwt) Freelist_intf.freelist_ops; >

val wrap : sync_origin:(('c, 'd) Freelist_intf.Fl_origin.t -> (unit, Tjr_monad.With_lwt.lwt) Tjr_monad.m) -> freelist_ops:('c, 'd, Tjr_monad.With_lwt.lwt) Freelist_intf.freelist_ops -> ('a -> ('b, Tjr_fs_shared.Shared_ctxt.t) Tjr_monad.m) -> 'a -> ('b, Tjr_monad.With_lwt.lwt) Tjr_monad.m

This wraps the freelist operations in an inefficient piece of code that checks whether the origin has changed, and if it has, it syncs the origin

val add_origin_autosync : origin_blk_id:Tjr_fs_shared.Shared_ctxt.B.blk_id -> (Tjr_fs_shared.Shared_ctxt.r, Tjr_fs_shared.Shared_ctxt.r, Tjr_monad.With_lwt.lwt) Freelist_intf.freelist_ops -> (Tjr_fs_shared.Shared_ctxt.r, Tjr_fs_shared.Shared_ctxt.r, Tjr_monad.With_lwt.lwt) Freelist_intf.freelist_ops

val initialize : origin:Tjr_fs_shared.Shared_ctxt.B.blk_id -> free_blk:Tjr_plist__.Pl_type_abbrevs.blk_id -> min_free:Tjr_fs_shared.Shared_ctxt.r option -> (unit, Tjr_monad.With_lwt.lwt) Tjr_monad.m

type fl_origin' = (Tjr_fs_shared.Shared_ctxt.r, Tjr_fs_shared.Shared_ctxt.r) Freelist_intf.Fl_origin.t

val fl_origin'_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> fl_origin'

val sexp_of_fl_origin' : fl_origin' -> Ppx_sexp_conv_lib.Sexp.t

val restore' : Tjr_fs_shared.Shared_ctxt.B.blk_id -> (< freelist_ops : (Tjr_fs_shared.Shared_ctxt.r, Tjr_fs_shared.Shared_ctxt.r, Tjr_monad.With_lwt.lwt) Freelist_intf.freelist_ops; freelist_ref : Tjr_fs_shared.Shared_ctxt.r Freelist_intf.freelist_im Stdlib.ref; >, Tjr_monad.With_lwt.lwt) Tjr_monad.m

val restore : autosync:bool -> origin:Tjr_fs_shared.Shared_ctxt.B.blk_id -> (< freelist_ops : (Tjr_fs_shared.Shared_ctxt.r, Tjr_fs_shared.Shared_ctxt.r, Tjr_monad.With_lwt.lwt) Freelist_intf.freelist_ops; freelist_ref : Tjr_fs_shared.Shared_ctxt.r Freelist_intf.freelist_im Stdlib.ref; >, Tjr_monad.With_lwt.lwt) Tjr_monad.m